Refactor: Modularize feature selection pipeline
This commit refactors the feature selection pipeline into a modular system. It removes the monolithic `final_feature_selection_pipeline.py` and `final_feature_selection_pipeline_modular.py` files, consolidating the core logic into `feature_selection.py`. The `FinalFeatureSelectionComponent` and `FinalFeatureSelectionStep` are updated to use the new modular structure. This change improves code organization and maintainability.
