--- src/training/steps/pre_training/final_feature_selection_pipeline.py
@@ -1,4226 +1,0 @@
-#!/usr/bin/env python3
-"""
-Final Feature Selection Pipeline
-
-This module implements a comprehensive multi-stage feature selection pipeline
-that runs at the end of the market analysis pipeline, progressively reducing
-features from 120 → 100 → 80 → 60 using RandomForest and SHAP analysis.
-"""
-
-import pandas as pd
-import numpy as np
-from typing import Dict, List, Optional, Tuple, Any, Union
-from dataclasses import dataclass, field
-import time
-import json
-import math
-import os
-from pathlib import Path
-from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
-from sklearn.metrics import average_precision_score, balanced_accuracy_score
-from sklearn.model_selection import StratifiedKFold, KFold
-from sklearn.preprocessing import StandardScaler
-from sklearn.impute import SimpleImputer
-from sklearn.feature_selection import mutual_info_regression
-from scipy.stats import rankdata
-import joblib
-from collections import defaultdict
-from joblib import Parallel, delayed
-
-# Try to import SHAP, fallback if not available
-try:
-    import shap
-    SHAP_AVAILABLE = True
-except ImportError:
-    SHAP_AVAILABLE = False
-    shap = None
-
-# Try to import LightGBM for improved feature selection
-try:
-    import lightgbm as lgb
-    LIGHTGBM_AVAILABLE = True
-except ImportError:
-    LIGHTGBM_AVAILABLE = False
-    lgb = None
-
-# Try to import sklearn feature selection for RFE
-try:
-    from sklearn.feature_selection import RFE, RFECV
-    from sklearn.model_selection import StratifiedKFold
-    SKLEARN_FEATURE_SELECTION_AVAILABLE = True
-except ImportError:
-    SKLEARN_FEATURE_SELECTION_AVAILABLE = False
-
-# Import consolidated VectorBT utilities
-from .utils.vectorbt_utils import (
-    create_vectorbt_tools, VectorBTConfig, get_vectorbt_performance_stats,
-    VECTORBT_UTILS_AVAILABLE
-)
-
-# Import matrix operations and hardware utilities
-try:
-    from src.utils.matrix_operations import (
-        get_unified_matrix_operations,
-        correlation_matrix_gpu,
-        matrix_correlation_analysis,
-        batch_correlation_analysis,
-        optimize_dataframe,
-        get_batch_matrix_processor
-    )
-    from src.utils.hardware import (
-        get_unified_hardware_manager,
-        get_adaptive_optimization_engine,
-        get_advanced_memory_optimizer,
-        WorkloadType
-    )
-    from src.utils.hardware.advanced_memory_optimizer import AdvancedM1MemoryOptimizer, MemoryStrategy
-    MATRIX_OPERATIONS_AVAILABLE = True
-    HARDWARE_OPTIMIZATION_AVAILABLE = True
-except ImportError:
-    MATRIX_OPERATIONS_AVAILABLE = False
-    HARDWARE_OPTIMIZATION_AVAILABLE = False
-
-# Import existing feature selection tools
-try:
-    from src.training.utils.feature_selection.selection_methods import (
-        mrmr_selection, lasso_selection, correlation_filtering,
-        recursive_feature_elimination, variance_filtering
-    )
-    from src.feature_selection.analysis.feature_importance_analyzer import (
-        FeatureImportanceAnalyzer, ImportanceMethod, FeatureImportanceConfig
-    )
-    from src.training.utils.feature_selection.quality_metrics import (
-        calculate_feature_quality_metrics, FeatureQualityMetrics
-    )
-    EXISTING_FEATURE_SELECTION_AVAILABLE = True
-except Exception:
-    EXISTING_FEATURE_SELECTION_AVAILABLE = False
-
-# Import system utilities
-from src.utils.logger import get_logger
-from src.utils.tprint import (
-    tprint, tprint_info, tprint_success, tprint_warning, tprint_error, tprint_debug
-)
-from src.feature_selection import EntropyBalancerConfig, EntropyFilterResult, EntropyStabilityFilter
-from .calibration_utils import compute_classification_calibration, evaluate_conformal_interval
-from .backtesting.turnover import (
-    apply_market_impact_model,
-    calculate_turnover_metrics,
-    reject_high_turnover_configs,
-)
-
-# Import gate feature protection
-try:
-    from .gate_feature_protection import GateFeatureProtector, GateFeatureConfig
-    GATE_PROTECTION_AVAILABLE = True
-except ImportError as e:
-    print(f"⚠️ WARNING: Gate feature protection not available: {e}")
-    GATE_PROTECTION_AVAILABLE = False
-
-try:
-    from src.utils.ml_common.validation.cv import PurgedKFoldTime
-    PURGED_KFOLD_AVAILABLE = PurgedKFoldTime is not None
-except Exception:
-    PurgedKFoldTime = None  # type: ignore
-    PURGED_KFOLD_AVAILABLE = False
-
-@dataclass
-class BaseFeatureSelectionConfig:
-    """Base configuration for multi-stage feature selection."""
-    # Stage targets
-    initial_features: int = 120
-    stage_1_target: int = 100
-    stage_2_target: int = 80
-    stage_3_target: int = 60
-
-    # Model configuration
-    model_type: str = 'regime_detection'
-    target_features: int = 80
-    min_features: int = 60
-    max_features: int = 100
-    priority_categories: List[str] = field(default_factory=lambda: ['volatility', 'structural', 'volume_regime', 'statistical'])
-
-    # VectorBT optimization settings
-    enable_vectorbt_optimization: bool = True
-    vectorbt_memory_efficient: bool = True
-    vectorbt_chunk_size: int = 1000
-    vectorbt_enable_parallel: bool = True
-    vectorbt_enable_gpu: bool = False
-
-    # Output settings
-    save_models: bool = True
-    save_analysis: bool = True
-    output_directory: str = "outcomes/market_analysis"
-    verbose: bool = True
-
-
-@dataclass
-class ModelSpecificConfig:
-    """Model-specific parameters."""
-    # RandomForest parameters
-    rf_n_estimators: int = 100
-    rf_max_depth: int = 10
-    rf_min_samples_split: int = 5
-    rf_random_state: int = 42
-
-    # SHAP parameters
-    shap_sample_size: int = 1000
-    shap_max_features: int = 200
-
-    # LightGBM parameters
-    lightgbm_params: Dict[str, Any] = field(default_factory=lambda: {
-        'n_estimators': 100,
-        'learning_rate': 0.1,
-        'max_depth': 10,
-        'num_leaves': 31,
-        'min_child_samples': 20,
-        'subsample': 0.8,
-        'colsample_bytree': 0.8,
-        'reg_alpha': 0.1,
-        'reg_lambda': 0.1,
-        'random_state': 42
-    })
-
-
-@dataclass
-class QualityThresholdsConfig:
-    """Quality thresholds for feature selection."""
-    min_feature_importance: float = 0.002
-    min_correlation_threshold: float = 0.90
-    min_variance_threshold: float = 0.005
-    model_correlation_threshold: float = 0.85
-    model_importance_threshold: float = 0.003
-    
-    # Regime-specific thresholds
-    regime_importance_threshold: float = 0.005
-    regime_correlation_threshold: float = 0.80
-    regime_variance_threshold: float = 0.001
-
-
-@dataclass
-class ValidationConfig:
-    """Cross-validation and trading-aware evaluation."""
-    cv_folds: int = 5
-    cv_scoring: str = 'neg_mean_squared_error'
-    label_horizon_minutes: int = 30
-    purge_minutes: Optional[int] = None
-    embargo_minutes: Optional[int] = None
-    
-    # Trading parameters
-    trading_cost: float = 0.0005
-    trading_horizon: int = 252
-    turnover_penalty: float = 0.0
-    ic_method: str = 'spearman'
-    market_impact_coefficient: float = 0.1
-    capacity_limit_usd: float = 1_000_000.0
-    max_turnover_annual: float = 50.0
-    min_sharpe_to_turnover_ratio: float = 0.1
-
-    # Uncertainty & calibration
-    enable_uncertainty_reporting: bool = True
-    reliability_bins: int = 10
-    confidence_coverage_target: float = 0.9
-
-
-@dataclass
-class AdvancedSelectionConfig:
-    """Advanced feature selection parameters."""
-    # Selection methods
-    selection_methods: List[str] = field(default_factory=lambda: [
-        'mrmr', 'lasso', 'correlation_filtering', 'rfe', 'variance_filtering', 'mutual_info'
-    ])
-
-    # Directional feature selection
-    direction_mode: str = 'both'
-    separate_directional_features: bool = True
-    directional_feature_prefixes: Dict[str, str] = field(default_factory=lambda: {
-        'long': 'long_',
-        'short': 'short_'
-    })
-
-    # Regime-aware selection
-    enable_regime_aware_selection: bool = True
-    regime_clustering_threshold: float = 0.7
-    regime_separation_bonus: float = 0.1
-    regime_focus_weights: Dict[str, float] = field(default_factory=lambda: {
-        'volatility': 0.35,
-        'structural': 0.25,
-        'volume_regime': 0.20,
-        'statistical': 0.20
-    })
-
-    # Multi-criteria selection
-    enable_multi_criteria_selection: bool = True
-    criteria_weights: Dict[str, float] = field(default_factory=lambda: {
-        'importance': 0.30,
-        'correlation': 0.20,
-        'variance': 0.15,
-        'regime_separation': 0.25,
-        'temporal_stability': 0.10
-    })
-    
-    # Stage-specific scoring weights
-    stage_1_weights: Dict[str, float] = field(default_factory=lambda: {
-        'mrmr': 0.7,
-        'spearman': 0.3
-    })
-    stage_3_weights: Dict[str, float] = field(default_factory=lambda: {
-        'ensemble': 0.6,
-        'stability': 0.4
-    })
-
-    # Entropy stability filtering
-    enable_entropy_balancing: bool = True
-    entropy_num_slices: int = 12
-    entropy_min_slice_size: int = 100
-    entropy_variance_threshold: float = 0.12
-    entropy_max_bins: int = 15
-    entropy_min_unique_values: int = 5
-    entropy_use_time_index: bool = True
-
-
-    # RFE parameters
-    enable_rfe: bool = True
-    rfe_step_size: float = 0.1
-    rfe_min_features: int = 10
-    rfe_cv_folds: int = 3
-    rfe_early_stopping: bool = True
-    rfe_early_stopping_patience: int = 3
-
-
-    # Chunked processing
-    enable_chunked_processing: bool = True
-    chunk_size: int = 1000
-    max_chunks: int = 10
-    chunk_overlap: int = 50
-
-
-
-@dataclass
-class FeatureSelectionConfig(BaseFeatureSelectionConfig):
-    """Complete configuration combining all sub-configs."""
-    
-    def __post_init__(self):
-        """Initialize sub-configurations."""
-        tprint_debug("🔧 Initializing feature selection configuration sub-configs")
-        self.model_config = ModelSpecificConfig()
-        self.quality_config = QualityThresholdsConfig()
-        self.validation_config = ValidationConfig()
-        self.advanced_config = AdvancedSelectionConfig()
-        
-        # Merge parameters from sub-configs for backward compatibility
-        for config in [self.model_config, self.quality_config, self.validation_config, self.advanced_config]:
-            for attr, value in config.__dict__.items():
-                if not hasattr(self, attr):
-                    setattr(self, attr, value)
-
-@dataclass
-class FeatureSelectionResult:
-    """Result of feature selection analysis."""
-    # Stage results
-    stage_1_features: List[str] = field(default_factory=list)
-    stage_2_features: List[str] = field(default_factory=list)
-    stage_3_features: List[str] = field(default_factory=list)
-    final_features: List[str] = field(default_factory=list)
-    
-    # Scores and metrics
-    stage_1_scores: Dict[str, Any] = field(default_factory=dict)
-    stage_2_scores: Dict[str, Any] = field(default_factory=dict)
-    stage_3_scores: Dict[str, Any] = field(default_factory=dict)
-    final_scores: Dict[str, Any] = field(default_factory=dict)
-    eligible_for_selection: bool = True
-    turnover_rejection_reason: Optional[str] = None
-    
-    # Feature importance
-    rf_importance: Dict[str, float] = field(default_factory=dict)
-    shap_importance: Dict[str, float] = field(default_factory=dict)
-    combined_importance: Dict[str, float] = field(default_factory=dict)
-    
-    # Analysis metadata
-    feature_counts: Dict[str, int] = field(default_factory=dict)
-    selection_time: float = 0.0
-    model_performance: Dict[str, Any] = field(default_factory=dict)
-    hypothesis_report: Dict[str, Any] = field(default_factory=dict)
-    horizon_p_values: Dict[str, float] = field(default_factory=dict)
-    feature_p_values: Dict[str, float] = field(default_factory=dict)
-    lookback_p_values: Dict[str, float] = field(default_factory=dict)
-    adjusted_p_values: Dict[str, Dict[str, float]] = field(default_factory=dict)
-    
-    # Quality metrics
-    correlation_analysis: Dict[str, Any] = field(default_factory=dict)
-    variance_analysis: Dict[str, Any] = field(default_factory=dict)
-    stability_scores: Dict[str, float] = field(default_factory=dict)
-    entropy_variance: Dict[str, float] = field(default_factory=dict)
-    entropy_stability: Dict[str, float] = field(default_factory=dict)
-    entropy_removed_features: Dict[str, float] = field(default_factory=dict)
-    calibration_metrics: Dict[str, Any] = field(default_factory=dict)
-    reliability_diagram: Dict[str, Any] = field(default_factory=dict)
-    uncertainty_metrics: Dict[str, Any] = field(default_factory=dict)
-
-class MultiStageFeatureSelector:
-    """Multi-stage feature selection using RandomForest and SHAP with vectorization and caching."""
-
-    def __init__(self, config: Optional[FeatureSelectionConfig] = None, execution_mode_config: Optional[Dict[str, Any]] = None):
-        self.config = config or FeatureSelectionConfig()
-        self.logger = get_logger("MultiStageFeatureSelector")
-        self.matrix_ops = get_unified_matrix_operations()
-
-        # Initialize VectorBT optimization tools if available
-        if VECTORBT_UTILS_AVAILABLE and self.config.enable_vectorbt_optimization:
-            tprint("🚀 Initializing VectorBT optimization tools")
-            vectorbt_config = VectorBTConfig(
-                enable_gpu=self.config.vectorbt_enable_gpu,
-                enable_parallel=self.config.vectorbt_enable_parallel,
-                memory_efficient=self.config.vectorbt_memory_efficient,
-                chunk_size=self.config.vectorbt_chunk_size
-            )
-            
-            vectorbt_tools = create_vectorbt_tools(vectorbt_config)
-            self.vectorbt_optimizer = vectorbt_tools['optimizer']
-            self.vectorization_manager = vectorbt_tools['manager']
-            self.vectorbt_enabled = vectorbt_tools['available']
-            
-            if self.vectorbt_enabled:
-                tprint("✅ VectorBT optimization tools initialized successfully")
-            else:
-                tprint_warning("⚠️ VectorBT optimization tools not available")
-        else:
-            self.vectorbt_optimizer = None
-            self.vectorization_manager = None
-            self.vectorbt_enabled = False
-            tprint("⚠️ VectorBT optimization disabled or not available")
-
-        # Initialize gate feature protection if available
-        if GATE_PROTECTION_AVAILABLE:
-            gate_config = self.config.custom_params.get('gate_protection', {})
-            if gate_config and isinstance(gate_config, dict):
-                self.gate_protector = GateFeatureProtector(GateFeatureConfig(**gate_config))
-            else:
-                self.gate_protector = GateFeatureProtector()
-            tprint("🛡️ Gate feature protection enabled")
-        else:
-            self.gate_protector = None
-
-        # Initialize hardware optimization tools if available
-        if HARDWARE_OPTIMIZATION_AVAILABLE:
-            self.hardware_manager = get_unified_hardware_manager()
-            self.adaptive_engine = get_adaptive_optimization_engine()
-            self.memory_optimizer = get_advanced_memory_optimizer()
-            
-            # Initialize advanced memory optimizer for aggressive cleanup
-            try:
-                self.advanced_memory_optimizer = AdvancedM1MemoryOptimizer(
-                    memory_limit_gb=8.0,
-                    strategy=MemoryStrategy.AGGRESSIVE
-                )
-                self.logger.info("🧠 Advanced memory optimizer initialized with aggressive strategy")
-            except Exception as e:
-                self.advanced_memory_optimizer = None
-                self.logger.warning(f"⚠️ Advanced memory optimizer not available: {e}")
-            
-            self.logger.info("🚀 Hardware optimization tools initialized")
-        else:
-            self.hardware_manager = None
-            self.adaptive_engine = None
-            self.memory_optimizer = None
-            self.advanced_memory_optimizer = None
-            self.logger.info("⚠️ Hardware optimization tools not available")
-
-        # Initialize batch processor for chunked processing
-        if MATRIX_OPERATIONS_AVAILABLE:
-            self.batch_processor = get_batch_matrix_processor()
-            self.logger.info("📦 Batch matrix processor initialized")
-        else:
-            self.batch_processor = None
-            self.logger.info("⚠️ Batch matrix processor not available")
-
-        # Initialize execution mode configuration
-        self.execution_mode_config = execution_mode_config
-        
-        # Initialize VectorBT memory optimization if available
-        if VECTORBT_AVAILABLE:
-            self._vectorbt_memory_config = {
-                'use_memory_efficient': True,
-                'chunk_size': 1000,
-                'max_memory_gb': 8.0,
-                'enable_gpu': False  # Can be enabled if GPU is available
-            }
-            tprint("🚀 VectorBT memory optimization enabled")
-        else:
-            self._vectorbt_memory_config = None
-            tprint("⚠️ VectorBT memory optimization not available")
-    
-    def _display_vectorbt_status(self):
-        """Display VectorBT optimization status and configuration."""
-        if VECTORBT_AVAILABLE:
-            tprint("🚀 VECTORBT OPTIMIZATION STATUS:")
-            tprint(f"   ✅ VectorBT Available: {VECTORBT_AVAILABLE}")
-            tprint(f"   🧠 Memory Optimization: {'Enabled' if self._vectorbt_memory_config else 'Disabled'}")
-            if self._vectorbt_memory_config:
-                tprint(f"   📊 Chunk Size: {self._vectorbt_memory_config.get('chunk_size', 'N/A')}")
-                tprint(f"   💾 Max Memory: {self._vectorbt_memory_config.get('max_memory_gb', 'N/A')}GB")
-                tprint(f"   🖥️ GPU Support: {'Enabled' if self._vectorbt_memory_config.get('enable_gpu', False) else 'Disabled'}")
-            tprint("   🔧 Optimized Operations:")
-            tprint("      - Correlation Analysis (large feature sets)")
-            tprint("      - Feature Importance Calculation (large feature sets)")
-            tprint("      - Ensemble Scoring (large feature sets)")
-            tprint("      - Stability Selection (large feature sets)")
-            tprint("      - mRMR Calculation (large feature sets)")
-            tprint("      - Memory Optimization (all stages)")
-        else:
-            tprint("⚠️ VECTORBT OPTIMIZATION STATUS:")
-            tprint("   ❌ VectorBT Not Available - using traditional methods")
-            tprint("   📊 Fallback to numpy/scikit-learn operations")
-        if self.execution_mode_config:
-            self.logger.info(f"📊 Using execution mode configuration for feature selection")
-        else:
-            self.logger.info("📊 No execution mode configuration provided, using defaults")
-
-        # Initialize results
-        self.results = FeatureSelectionResult()
-        self.results.polarity_adjustments = {}
-        self.results.sign_stability = {}
-        self._active_event_timestamps: Optional[pd.Series] = None
-
-        # Set model-specific parameters
-        self._set_model_specific_parameters()
-
-        # Initialize directional results for separate long/short feature sets
-        if self.config.separate_directional_features:
-            self.long_results = FeatureSelectionResult()
-            self.short_results = FeatureSelectionResult()
-            self.long_results.polarity_adjustments = {}
-            self.short_results.polarity_adjustments = {}
-            self.long_results.sign_stability = {}
-            self.short_results.sign_stability = {}
-
-        # Initialize thread-safe caching system for vectorized operations
-        import threading
-        self._cache_lock = threading.RLock()
-        self._cache = {}
-        self._cache_hits = 0
-        self._cache_misses = 0
-
-        # Initialize vectorized computation arrays
-        self._vectorized_arrays = {}
-        self._computation_cache = {}
-
-        # Initialize polarity tracking containers
-        self.feature_polarity_adjustments: Dict[str, Dict[str, Any]] = {}
-        self.feature_polarity_history: Dict[str, List[float]] = {}
-        self.feature_sign_stability: Dict[str, float] = {}
-
-        tprint("🚀 MultiStageFeatureSelector initialized with vectorization and caching")
-        tprint(f"🎯 Model Type: {self.config.model_type}")
-        tprint(f"📊 Feature Range: {self.config.min_features}-{self.config.max_features} (target: {self.config.target_features})")
-        tprint(f"🎯 Direction Mode: {self.config.direction_mode}")
-        tprint(f"🎯 Separate Features: {self.config.separate_directional_features}")
-        tprint(f"⚡ LightGBM: {LIGHTGBM_AVAILABLE}")
-        tprint(f"⚡ RFE: {self.config.enable_rfe}")
-        tprint(f"⚡ Chunked Processing: {self.config.enable_chunked_processing}")
-        tprint(f"⚡ Vectorization: Enabled")
-        tprint(f"⚡ Caching: Enabled")
-        
-        # Memory management thresholds - optimized for M1
-        self.cache_size_limit_mb = 200  # 200MB cache limit for M1
-        self.max_cache_entries = 500    # Maximum number of cache entries
-        self.memory_pressure_threshold = 0.8  # Trigger cleanup at 80%
-        self.aggressive_cleanup_threshold = 0.9  # Aggressive cleanup at 90%
-        
-        # Initialize caches and state for vectorized operations
-        self._variance_cache = {}
-        self._X_cached = None
-        self._last_X = None
-        
-        # Initialize random state for bootstrap variation
-        self._rng = np.random.default_rng(self.config.random_state)
-        
-        # Set thread limits to avoid oversubscription
-        self._set_thread_limits()
-
-    def _set_thread_limits(self):
-        """Set thread limits to avoid oversubscription."""
-        tprint_debug("🔧 Setting thread limits to avoid oversubscription")
-        # Set MKL and OpenMP threads
-        os.environ['MKL_NUM_THREADS'] = str(min(4, os.cpu_count() // 2))
-        os.environ['OMP_NUM_THREADS'] = str(min(4, os.cpu_count() // 2))
-        
-        # LightGBM thread limit
-        self.lgb_threads = min(4, os.cpu_count() // 2)
-
-    def _y_numeric(self, y: pd.Series) -> np.ndarray:
-        """Convert y to numeric for Spearman/ranks."""
-        if pd.api.types.is_numeric_dtype(y):
-            return y.to_numpy()
-        return pd.Categorical(y).codes.astype(float)
-
-    def spearman_abs_vectorized(self, X: pd.DataFrame, y: pd.Series) -> pd.Series:
-        """Optimized vectorized Spearman correlation calculation with early termination."""
-        # Fast fail on empty data
-        if X.empty or y.empty:
-            return pd.Series(dtype=float, index=X.columns)
-        
-        # Use cached computation if available
-        cache_key = f"spearman_{hash(str(X.columns.tolist()))}_{hash(str(y.name))}"
-        cached_result = self._get_from_cache(cache_key)
-        if cached_result is not None:
-            return cached_result
-        
-        try:
-            # Optimized ranking with early termination check
-            Xr = X.rank(method="average", na_option='keep').to_numpy(dtype=float)
-            yr = rankdata(self._y_numeric(y))
-            
-            # Early termination if all features have low variance
-            X_var = np.var(Xr, axis=0)
-            if np.max(X_var) < 1e-8:
-                tprint_warning("⚠️ All features have very low variance, returning zero correlations")
-                result = pd.Series(0.0, index=X.columns)
-                self._set_cache(cache_key, result)
-                return result
-            
-            # Center with optimized operations
-            X_mean = np.mean(Xr, axis=0, keepdims=True)
-            y_mean = np.mean(yr)
-            
-            Xc = Xr - X_mean
-            yc = yr - y_mean
-            
-            # Compute correlations with numerical stability
-            Xc_norm = np.sqrt(np.sum(Xc**2, axis=0))
-            yc_norm = np.sqrt(np.sum(yc**2))
-            
-            # Avoid division by zero
-            Xc_norm = np.where(Xc_norm < 1e-8, 1.0, Xc_norm)
-            yc_norm = max(yc_norm, 1e-8)
-            
-            # Vectorized correlation computation
-            corrs = (Xc.T @ yc) / (Xc_norm * yc_norm)
-            
-            # Early termination if correlations are too low
-            max_corr = np.max(np.abs(corrs))
-            if max_corr < 0.01:
-                tprint_warning("⚠️ All correlations are very low, early termination")
-                result = pd.Series(0.0, index=X.columns)
-                self._set_cache(cache_key, result)
-                return result
-            
-            result = pd.Series(np.abs(corrs), index=X.columns)
-            self._set_cache(cache_key, result)
-            return result
-            
-        except Exception as e:
-            tprint_error(f"❌ Spearman correlation calculation failed: {e}")
-            # Return zero correlations on error
-            result = pd.Series(0.0, index=X.columns)
-            self._set_cache(cache_key, result)
-            return result
-
-    def redundancy_mean_abs_spearman_blocked(self, X: pd.DataFrame, block=1024) -> pd.Series:
-        """Blocked redundancy calculation for large feature sets."""
-        R = X.rank(method="average")
-        p = R.shape[1]
-        sums = np.zeros(p, dtype=float)
-        counts = np.zeros(p, dtype=float)
-        
-        for i in range(0, p, block):
-            Ri = R.iloc[:, i:i+block]
-            Ci = Ri.corr(method="pearson").abs().to_numpy()
-            np.fill_diagonal(Ci, np.nan)
-            sums[i:i+block] += np.nansum(Ci, axis=1)
-            counts[i:i+block] += np.sum(~np.isnan(Ci), axis=1)
-        
-        red = sums / np.maximum(counts, 1)
-        return pd.Series(red, index=X.columns)
-
-    def zscore_matrix(self, M: np.ndarray, axis=0) -> np.ndarray:
-        """Vectorized z-score normalization."""
-        mu = M.mean(axis=axis, keepdims=True)
-        sd = M.std(axis=axis, ddof=0, keepdims=True)
-        return (M - mu) / (sd + 1e-12)
-
-    def top_k(self, names: List[str], scores: np.ndarray, k: int) -> List[str]:
-        """Top-k selection using argpartition for O(p) performance."""
-        k = min(k, len(scores))
-        idx = np.argpartition(-scores, k-1)[:k]   # top k unordered
-        
-        # Order those k deterministically
-        sub_scores = scores[idx]
-        sub_names = np.array(names)[idx]
-        order = np.lexsort((sub_names, -sub_scores))  # desc by score, then name asc
-        sub = idx[order]
-        
-        return [names[i] for i in sub]
-
-    def bottom_k(self, names: List[str], scores: np.ndarray, k: int) -> List[str]:
-        """Bottom-k selection using argpartition for O(p) performance."""
-        k = min(k, len(scores))
-        idx = np.argpartition(scores, k-1)[:k]
-        
-        # Order those k deterministically
-        sub_scores = scores[idx]
-        sub_names = np.array(names)[idx]
-        order = np.lexsort((sub_names, sub_scores))  # asc by score, then name asc
-        sub = idx[order]
-        
-        return [names[i] for i in sub]
-
-    def _vectorbt_optimized_correlation_matrix(self, X: pd.DataFrame) -> np.ndarray:
-        """Calculate correlation matrix using VectorBT optimization."""
-        if not self.vectorbt_enabled or self.vectorization_manager is None:
-            # Fallback to standard correlation
-            return X.corr().values
-        
-        try:
-            # Use VectorBT rolling correlation for optimized computation
-            correlation_result = self.vectorization_manager.rolling_operation(
-                X, 'corr', window=len(X), other=X
-            )
-            return correlation_result.values
-        except Exception as e:
-            self.logger.warning(f"VectorBT correlation failed: {e}, using fallback")
-            return X.corr().values
-
-    def _vectorbt_optimized_rolling_operations(self, X: pd.DataFrame, operation: str, window: int = 20) -> pd.DataFrame:
-        """Perform rolling operations using VectorBT optimization."""
-        if not self.vectorbt_enabled or self.vectorization_manager is None:
-            # Fallback to pandas rolling
-            return getattr(X.rolling(window=window), operation)()
-        
-        try:
-            # Use VectorBT rolling optimizer
-            result = self.vectorization_manager.rolling_operation(X, operation, window)
-            return result
-        except Exception as e:
-            self.logger.warning(f"VectorBT rolling {operation} failed: {e}, using fallback")
-            return getattr(X.rolling(window=window), operation)()
-
-    def _vectorbt_optimized_scaling(self, X: pd.DataFrame, method: str = 'zscore') -> pd.DataFrame:
-        """Scale data using VectorBT optimization."""
-        if not self.vectorbt_enabled or self.vectorization_manager is None:
-            # Fallback to standard scaling
-            if method == 'zscore':
-                return (X - X.mean()) / X.std()
-            elif method == 'minmax':
-                return (X - X.min()) / (X.max() - X.min())
-            else:
-                return X
-        
-        try:
-            # Use VectorBT scaling
-            result = self.vectorization_manager.scale_data(X, method)
-            return result
-        except Exception as e:
-            self.logger.warning(f"VectorBT scaling failed: {e}, using fallback")
-            if method == 'zscore':
-                return (X - X.mean()) / X.std()
-            elif method == 'minmax':
-                return (X - X.min()) / (X.max() - X.min())
-            else:
-                return X
-
-    def _vectorbt_optimized_batch_processing(self, feature_configs: List[Dict[str, Any]], data: pd.DataFrame) -> pd.DataFrame:
-        """Process multiple features in batch using VectorBT optimization."""
-        if not self.vectorbt_enabled or self.vectorization_manager is None:
-            # Fallback to sequential processing
-            results = {}
-            for config in feature_configs:
-                feature_name = config['name']
-                feature_type = config.get('type', 'rolling')
-                params = config.get('params', {})
-                
-                if feature_type == 'rolling':
-                    operation = params.get('operation', 'mean')
-                    window = params.get('window', 20)
-                    column = params.get('column', 'close')
-                    
-                    if column in data.columns:
-                        results[feature_name] = getattr(data[column].rolling(window=window), operation)()
-                
-                elif feature_type == 'scaling':
-                    method = params.get('method', 'zscore')
-                    column = params.get('column', 'close')
-                    
-                    if column in data.columns:
-                        if method == 'zscore':
-                            results[feature_name] = (data[column] - data[column].mean()) / data[column].std()
-                        elif method == 'minmax':
-                            results[feature_name] = (data[column] - data[column].min()) / (data[column].max() - data[column].min())
-            
-            return pd.DataFrame(results, index=data.index)
-        
-        try:
-            # Use VectorBT batch processing
-            result = self.vectorization_manager.batch_process_features(data, feature_configs)
-            return result
-        except Exception as e:
-            self.logger.warning(f"VectorBT batch processing failed: {e}, using fallback")
-            # Fallback to sequential processing
-            results = {}
-            for config in feature_configs:
-                feature_name = config['name']
-                feature_type = config.get('type', 'rolling')
-                params = config.get('params', {})
-                
-                if feature_type == 'rolling':
-                    operation = params.get('operation', 'mean')
-                    window = params.get('window', 20)
-                    column = params.get('column', 'close')
-                    
-                    if column in data.columns:
-                        results[feature_name] = getattr(data[column].rolling(window=window), operation)()
-                
-                elif feature_type == 'scaling':
-                    method = params.get('method', 'zscore')
-                    column = params.get('column', 'close')
-                    
-                    if column in data.columns:
-                        if method == 'zscore':
-                            results[feature_name] = (data[column] - data[column].mean()) / data[column].std()
-                        elif method == 'minmax':
-                            results[feature_name] = (data[column] - data[column].min()) / (data[column].max() - data[column].min())
-            
-            return pd.DataFrame(results, index=data.index)
-
-    def _lasso_scores_fast(self, X: pd.DataFrame, y: pd.Series, rs: int, n_splits: int = 3) -> pd.Series:
-        """Fast LASSO scoring with imputation."""
-        # Impute missing values
-        imp = SimpleImputer(strategy="median")
-        X_imp = imp.fit_transform(X)
-        
-        # Standardize
-        scaler = StandardScaler()
-        X_scaled = scaler.fit_transform(X_imp)
-        
-        # LASSO with warm start
-        lasso = LassoCV(cv=n_splits, random_state=rs, n_jobs=1)
-        lasso.fit(X_scaled, y)
-        
-        return pd.Series(np.abs(lasso.coef_), index=X.columns)
-
-    def _lgbm_shap_fast(self, X: pd.DataFrame, y: pd.Series, rs: int, task: str = "reg") -> pd.Series:
-        """Fast LightGBM with built-in TreeSHAP."""
-        # Configure LightGBM
-        if task == "reg":
-            lgb_model = lgb.LGBMRegressor(
-                n_estimators=200,
-                importance_type="gain",
-                random_state=rs,
-                verbose=-1,
-                num_threads=self.lgb_threads
-            )
-        else:
-            lgb_model = lgb.LGBMClassifier(
-                n_estimators=200,
-                importance_type="gain",
-                random_state=rs,
-                verbose=-1,
-                num_threads=self.lgb_threads
-            )
-        
-        lgb_model.fit(X, y)
-        
-        # Use built-in TreeSHAP (C++ speed)
-        try:
-            shap_values = lgb_model.predict(X, pred_contrib=True)
-            # Sum absolute SHAP values across samples
-            shap_importance = np.abs(shap_values[:, :-1]).mean(axis=0)  # Exclude bias term
-        except Exception:
-            # Fallback to feature importance
-            shap_importance = lgb_model.feature_importances_
-        
-        return pd.Series(shap_importance, index=X.columns)
-
-    def ensemble_scores_cv_parallel(self, X: pd.DataFrame, y: pd.Series, rs: int, task: str = "reg", n_splits: int = 5) -> pd.Series:
-        """Parallel CV ensemble with VectorBT optimization and caching."""
-        # Create cache key for this specific calculation
-        cache_key = f"ensemble_{hash(str(X.columns.tolist()))}_{len(X)}_{rs}_{task}_{n_splits}"
-        
-        # Check cache first
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        
-        try:
-            if VECTORBT_AVAILABLE and X.shape[1] > 50:  # Use VectorBT for larger feature sets
-                tprint("🚀 Using VectorBT for ensemble scoring")
-                result = self._vectorbt_ensemble_scores(X, y, rs, task, n_splits)
-            else:
-                tprint("📊 Using traditional parallel ensemble scoring")
-                result = self._traditional_ensemble_scores(X, y, rs, task, n_splits)
-            
-            # Cache the result
-            self._cache[cache_key] = result
-            return result
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ Ensemble scoring failed: {e}")
-            # Fallback to traditional method
-            return self._traditional_ensemble_scores(X, y, rs, task, n_splits)
-    
-    def _vectorbt_ensemble_scores(self, X: pd.DataFrame, y: pd.Series, rs: int, task: str = "reg", n_splits: int = 5) -> pd.Series:
-        """VectorBT-optimized ensemble scoring."""
-        try:
-            cols = X.columns.to_numpy()
-            p = len(cols)
-            
-            # Use VectorBT's vectorized operations for faster computation
-            kf = KFold(n_splits=n_splits, shuffle=True, random_state=rs)
-            
-            def fold_scores_vectorbt(tr):
-                Xtr, ytr = X.iloc[tr], y.iloc[tr]
-                
-                # Use VectorBT for faster LASSO scoring
-                s_lasso = self._vectorbt_lasso_scores(Xtr, ytr, rs)
-                
-                # Use VectorBT for faster LightGBM scoring
-                s_lgb = self._vectorbt_lgbm_scores(Xtr, ytr, rs, task)
-                
-                # Use VectorBT for z-score normalization
-                M = np.vstack([s_lasso.values, s_lgb.values]).T
-                Mz = vbt.rolling_apply(M, lambda x: (x - np.mean(x, axis=0)) / (np.std(x, axis=0) + 1e-8), window=len(x), step=1)
-                return Mz.mean(axis=1)
-            
-            # Process folds with VectorBT optimization
-            fold_vecs = []
-            for tr, _ in kf.split(X):
-                fold_vecs.append(fold_scores_vectorbt(tr))
-            
-            S = np.vstack(fold_vecs).mean(axis=0)
-            return pd.Series(S, index=cols)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT ensemble scoring failed: {e}")
-            return self._traditional_ensemble_scores(X, y, rs, task, n_splits)
-    
-    def _traditional_ensemble_scores(self, X: pd.DataFrame, y: pd.Series, rs: int, task: str = "reg", n_splits: int = 5) -> pd.Series:
-        """Traditional ensemble scoring method."""
-        kf = KFold(n_splits=n_splits, shuffle=True, random_state=rs)
-        cols = X.columns.to_numpy()
-        p = len(cols)
-
-        def fold_scores(tr):
-            Xtr, ytr = X.iloc[tr], y.iloc[tr]
-            
-            # Fast LASSO
-            s_lasso = self._lasso_scores_fast(Xtr, ytr, rs, n_splits=3)
-            
-            # Fast LightGBM with TreeSHAP
-            s_lgb = self._lgbm_shap_fast(Xtr, ytr, rs, task)
-            
-            # Combine scores with z-score normalization
-            M = np.vstack([s_lasso.values, s_lgb.values]).T     # (p, 2)
-            Mz = self.zscore_matrix(M, axis=0)                  # z-score by column
-            return Mz.mean(axis=1)                              # (p,)
-
-        # Parallel fold processing
-        fold_vecs = Parallel(n_jobs=-1, prefer="threads")(
-            delayed(fold_scores)(tr) for tr, _ in kf.split(X)
-        )
-        
-        S = np.vstack(fold_vecs).mean(axis=0)                   # (p,)
-        return pd.Series(S, index=cols)
-    
-    def _vectorbt_lasso_scores(self, X: pd.DataFrame, y: pd.Series, rs: int) -> pd.Series:
-        """VectorBT-optimized LASSO scoring."""
-        try:
-            from sklearn.linear_model import LassoCV
-            from sklearn.preprocessing import StandardScaler
-            
-            # Use VectorBT for faster preprocessing
-            scaler = StandardScaler()
-            X_scaled = scaler.fit_transform(X)
-            
-            # Use VectorBT's vectorized operations for LASSO
-            lasso = LassoCV(cv=3, random_state=rs, n_jobs=-1)
-            lasso.fit(X_scaled, y)
-            
-            # Get coefficients as importance scores
-            scores = np.abs(lasso.coef_)
-            return pd.Series(scores, index=X.columns)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT LASSO scoring failed: {e}")
-            # Fallback to traditional method
-            return self._lasso_scores_fast(X, y, rs, n_splits=3)
-    
-    def _vectorbt_lgbm_scores(self, X: pd.DataFrame, y: pd.Series, rs: int, task: str = "reg") -> pd.Series:
-        """VectorBT-optimized LightGBM scoring."""
-        try:
-            if not LIGHTGBM_AVAILABLE:
-                # Fallback to traditional method
-                return self._lgbm_shap_fast(X, y, rs, task)
-            
-            # Use VectorBT for faster LightGBM training
-            if task == "reg":
-                lgb_model = lgb.LGBMRegressor(
-                    n_estimators=50,
-                    learning_rate=0.1,
-                    random_state=rs,
-                    n_jobs=-1,
-                    verbose=-1
-                )
-            else:
-                lgb_model = lgb.LGBMClassifier(
-                    n_estimators=50,
-                    learning_rate=0.1,
-                    random_state=rs,
-                    n_jobs=-1,
-                    verbose=-1
-                )
-            
-            lgb_model.fit(X, y)
-            
-            # Get feature importance
-            importance = lgb_model.feature_importances_
-            return pd.Series(importance, index=X.columns)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT LightGBM scoring failed: {e}")
-            # Fallback to traditional method
-            return self._lgbm_shap_fast(X, y, rs, task)
-
-    def stability_scores_vectorized(self, X: pd.DataFrame, y: pd.Series, scorer, rs: int, n_boot=15, row_frac=0.7) -> pd.Series:
-        """Vectorized stability selection with VectorBT optimization."""
-        n = len(X)
-        p = X.shape[1]
-        k = max(1, p // 2)
-        cols = X.columns.to_numpy()
-        
-        try:
-            if VECTORBT_AVAILABLE and p > 100:  # Use VectorBT for large feature sets
-                tprint("🚀 Using VectorBT for stability scoring")
-                return self._vectorbt_stability_scores(X, y, scorer, rs, n_boot, row_frac, k, cols)
-            else:
-                tprint("📊 Using traditional stability scoring")
-                return self._traditional_stability_scores(X, y, scorer, rs, n_boot, row_frac, k, cols)
-                
-        except Exception as e:
-            tprint_warning(f"⚠️ Stability scoring failed: {e}")
-            return self._traditional_stability_scores(X, y, scorer, rs, n_boot, row_frac, k, cols)
-    
-    def _vectorbt_stability_scores(self, X: pd.DataFrame, y: pd.Series, scorer, rs: int, n_boot: int, row_frac: float, k: int, cols: np.ndarray) -> pd.Series:
-        """VectorBT-optimized stability scoring."""
-        try:
-            n = len(X)
-            p = X.shape[1]
-            freq = np.zeros(p, dtype=float)
-            
-            # Use VectorBT's vectorized operations for bootstrap sampling
-            for b in range(n_boot):
-                idx = self._rng.choice(n, int(max(2, row_frac * n)), replace=False)
-                
-                # Use VectorBT for faster scoring
-                X_boot = X.iloc[idx]
-                y_boot = y.iloc[idx]
-                
-                # Apply scorer with VectorBT optimization
-                s = scorer(X_boot, y_boot).to_numpy()
-                
-                # Use VectorBT's argpartition for faster selection
-                keep_idx = np.argpartition(-s, k-1)[:k]
-                freq[keep_idx] += 1
-            
-            return pd.Series(freq / n_boot, index=cols)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT stability scoring failed: {e}")
-            return self._traditional_stability_scores(X, y, scorer, rs, n_boot, row_frac, k, cols)
-    
-    def _traditional_stability_scores(self, X: pd.DataFrame, y: pd.Series, scorer, rs: int, n_boot: int, row_frac: float, k: int, cols: np.ndarray) -> pd.Series:
-        """Traditional stability scoring method."""
-        n = len(X)
-        p = X.shape[1]
-        freq = np.zeros(p, dtype=float)
-        
-        for b in range(n_boot):
-            idx = self._rng.choice(n, int(max(2, row_frac * n)), replace=False)
-            s = scorer(X.iloc[idx], y.iloc[idx]).to_numpy()    # (p,)
-            keep_idx = np.argpartition(-s, k-1)[:k]
-            freq[keep_idx] += 1
-        
-        return pd.Series(freq / n_boot, index=cols)
-
-    def _calculate_mrmr_mid_vectorized(self, X: pd.DataFrame, y: pd.Series) -> pd.Series:
-        """Fast MID approximation using VectorBT-optimized operations."""
-        try:
-            if VECTORBT_AVAILABLE and X.shape[1] > 100:  # Use VectorBT for large feature sets
-                tprint("🚀 Using VectorBT for mRMR calculation")
-                return self._vectorbt_mrmr_calculation(X, y)
-            else:
-                tprint("📊 Using traditional mRMR calculation")
-                return self._traditional_mrmr_calculation(X, y)
-                
-        except Exception as e:
-            tprint_warning(f"⚠️ mRMR calculation failed: {e}")
-            return self._traditional_mrmr_calculation(X, y)
-    
-    def _vectorbt_mrmr_calculation(self, X: pd.DataFrame, y: pd.Series) -> pd.Series:
-        """VectorBT-optimized mRMR calculation."""
-        try:
-            # Handle constants safely
-            const_mask = X.nunique(dropna=False) <= 1
-            if const_mask.any():
-                X_clean = X.loc[:, ~const_mask]
-            else:
-                X_clean = X
-            
-            # Use VectorBT for faster MI calculation
-            mi = mutual_info_regression(
-                X_clean, y, 
-                discrete_features=False,
-                random_state=self.config.random_state,
-                n_neighbors=min(3, max(2, int(np.sqrt(len(X_clean)))))
-            )
-            mi_s = pd.Series(mi, index=X_clean.columns)
-            
-            # Use VectorBT for redundancy calculation
-            if len(X_clean.columns) > 2000:
-                red = self._vectorbt_redundancy_calculation(X_clean)
-            else:
-                red = self._calculate_incremental_redundancy(X_clean)
-            
-            # MID = MI / (Redundancy + epsilon) - mathematically correct mRMR
-            red_aligned = red.reindex(mi_s.index).fillna(0.0)
-            mid = mi_s / (red_aligned + 1e-8)  # Add small epsilon to avoid division by zero
-            
-            # Reintroduce dropped constants with very low score
-            if const_mask.any():
-                mid = mid.reindex(X.columns).fillna(mid.min() - 1.0)
-            
-            return mid.reindex(X.columns).fillna(mid.min() - 1.0)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT mRMR calculation failed: {e}")
-            return self._traditional_mrmr_calculation(X, y)
-    
-    def _traditional_mrmr_calculation(self, X: pd.DataFrame, y: pd.Series) -> pd.Series:
-        """Traditional mRMR calculation method."""
-        # Handle constants safely
-        const_mask = X.nunique(dropna=False) <= 1
-        if const_mask.any():
-            X_clean = X.loc[:, ~const_mask]
-        else:
-            X_clean = X
-        
-        # Relevance via MI (kNN)
-        mi = mutual_info_regression(
-            X_clean, y, 
-            discrete_features=False,
-            random_state=self.config.random_state,
-            n_neighbors=min(3, max(2, int(np.sqrt(len(X_clean)))))
-        )
-        mi_s = pd.Series(mi, index=X_clean.columns)
-        
-        # Redundancy: use incremental calculation to avoid full matrix recalculation
-        if len(X_clean.columns) > 2000:
-            red = self.redundancy_mean_abs_spearman_blocked(X_clean)
-        else:
-            # Use incremental correlation calculation for better performance
-            red = self._calculate_incremental_redundancy(X_clean)
-        
-        # MID = MI / (Redundancy + epsilon) - mathematically correct mRMR
-        red_aligned = red.reindex(mi_s.index).fillna(0.0)
-        mid = mi_s / (red_aligned + 1e-8)  # Add small epsilon to avoid division by zero
-        
-        # Reintroduce dropped constants with very low score
-        if const_mask.any():
-            mid = mid.reindex(X.columns).fillna(mid.min() - 1.0)
-        
-        return mid.reindex(X.columns).fillna(mid.min() - 1.0)
-    
-    def _vectorbt_redundancy_calculation(self, X: pd.DataFrame) -> pd.Series:
-        """VectorBT-optimized redundancy calculation."""
-        try:
-            # Use VectorBT's correlation matrix for faster redundancy calculation
-            corr_matrix = vbt.correlation_matrix(X.values, method='spearman')
-            corr_matrix = corr_matrix.values
-            
-            # Calculate mean absolute correlation for each feature
-            redundancy_scores = np.mean(np.abs(corr_matrix), axis=1)
-            
-            return pd.Series(redundancy_scores, index=X.columns)
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT redundancy calculation failed: {e}")
-            # Fallback to traditional method
-            return self.redundancy_mean_abs_spearman_blocked(X)
-    
-    def _vectorbt_memory_optimization(self, X: pd.DataFrame, operation_name: str = "feature_selection") -> pd.DataFrame:
-        """Apply VectorBT memory optimization to DataFrame."""
-        try:
-            if not VECTORBT_AVAILABLE or self._vectorbt_memory_config is None:
-                return X
-            
-            tprint(f"🧠 Applying VectorBT memory optimization for {operation_name}")
-            
-            # Use VectorBT's memory-efficient operations
-            if self._vectorbt_memory_config.get('use_memory_efficient', True):
-                # Convert to VectorBT format for memory optimization
-                vbt_data = vbt.Data(X.values, index=X.index, columns=X.columns)
-                
-                # Apply memory optimization
-                optimized_data = vbt_data.memory_optimize()
-                
-                # Convert back to DataFrame
-                optimized_df = pd.DataFrame(
-                    optimized_data.values,
-                    index=optimized_data.index,
-                    columns=optimized_data.columns
-                )
-                
-                tprint(f"✅ VectorBT memory optimization applied: {X.memory_usage(deep=True).sum() / 1024**2:.2f}MB → {optimized_df.memory_usage(deep=True).sum() / 1024**2:.2f}MB")
-                
-                return optimized_df
-            else:
-                return X
-                
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT memory optimization failed: {e}")
-            return X
-
-    def _calculate_incremental_redundancy(self, X: pd.DataFrame) -> pd.Series:
-        """Calculate redundancy using incremental correlation to avoid full matrix computation."""
-        cols = X.columns.tolist()
-        n_features = len(cols)
-        redundancy_scores = np.zeros(n_features)
-        
-        # Calculate mean absolute correlation for each feature incrementally
-        for i, feature in enumerate(cols):
-            feature_corrs = []
-            for j, other_feature in enumerate(cols):
-                if i != j:  # Skip self-correlation
-                    corr = X[feature].corr(X[other_feature], method='spearman')
-                    if not np.isnan(corr):
-                        feature_corrs.append(abs(corr))
-            
-            # Use mean of absolute correlations as redundancy score
-            if feature_corrs:
-                redundancy_scores[i] = np.mean(feature_corrs)
-            else:
-                redundancy_scores[i] = 0.0
-        
-        return pd.Series(redundancy_scores, index=cols)
-
-    def _cleanup_cache(self, threshold_mb: float = 500) -> None:
-        """Clear cache if memory usage exceeds threshold."""
-        import sys
-        
-        try:
-            # Calculate cache size
-            cache_size_mb = sum(sys.getsizeof(v) for v in self._cache.values()) / 1e6
-            
-            # Check memory pressure and adjust threshold dynamically
-            memory_pressure = self._get_memory_pressure()
-            if memory_pressure > 0.8:  # High memory pressure
-                threshold_mb = min(threshold_mb, 200)  # Lower threshold under pressure
-            elif memory_pressure > 0.6:  # Medium memory pressure
-                threshold_mb = min(threshold_mb, 300)  # Moderate threshold
-            
-            if cache_size_mb > threshold_mb or memory_pressure > 0.7:
-                # Clear caches
-                cleared_entries = len(self._cache)
-                self._cache.clear()
-                self._vectorized_arrays.clear()
-                self._computation_cache.clear()
-                
-                tprint(f"🧹 Cleared {cleared_entries} cache entries ({cache_size_mb:.1f}MB) - memory pressure: {memory_pressure:.2f}")
-                self.logger.info(f"Cache cleanup: cleared {cache_size_mb:.1f}MB, pressure: {memory_pressure:.2f}")
-                
-                # Reset cache statistics
-                self._cache_hits = 0
-                self._cache_misses = 0
-        except Exception as e:
-            tprint_warning(f"⚠️ Cache cleanup failed: {e}")
-            self.logger.warning(f"Cache cleanup error: {e}")
-    
-    def _aggressive_memory_cleanup(self, force_cleanup: bool = False) -> Dict[str, Any]:
-        """Perform aggressive memory cleanup using advanced memory optimizer."""
-        cleanup_results = {
-            'success': False,
-            'memory_freed_mb': 0.0,
-            'memory_pressure_before': 0.0,
-            'memory_pressure_after': 0.0,
-            'cleanup_methods_used': [],
-            'errors': []
-        }
-        
-        try:
-            # Get initial memory pressure
-            cleanup_results['memory_pressure_before'] = self._get_memory_pressure()
-            
-            # Use advanced memory optimizer if available
-            if hasattr(self, 'advanced_memory_optimizer') and self.advanced_memory_optimizer:
-                tprint("🧹 Performing aggressive memory cleanup with advanced optimizer...")
-                
-                # Perform comprehensive cleanup
-                advanced_cleanup = self.advanced_memory_optimizer.aggressive_cleanup(
-                    force_cleanup=force_cleanup,
-                    clear_caches=True,
-                    compress_memory=True,
-                    optimize_pools=True
-                )
-                
-                cleanup_results.update({
-                    'success': advanced_cleanup.get('success', False),
-                    'memory_freed_mb': advanced_cleanup.get('memory_freed_mb', 0.0),
-                    'cleanup_methods_used': advanced_cleanup.get('methods_used', [])
-                })
-                
-                self.logger.info(f"🧹 Advanced memory cleanup: {cleanup_results['memory_freed_mb']:.1f}MB freed")
-            
-            # Fallback to standard cleanup
-            if not cleanup_results['success']:
-                tprint("🧹 Performing fallback memory cleanup...")
-                
-                # Clear caches
-                self._clear_cache()
-                self._vectorized_arrays.clear()
-                self._computation_cache.clear()
-                
-                # Force garbage collection
-                import gc
-                collected = gc.collect()
-                
-                cleanup_results.update({
-                    'success': True,
-                    'memory_freed_mb': collected * 0.001,  # Rough estimate
-                    'cleanup_methods_used': ['cache_clear', 'garbage_collection']
-                })
-                
-                self.logger.info(f"🧹 Fallback memory cleanup: {collected} objects collected")
-            
-            # Get final memory pressure
-            cleanup_results['memory_pressure_after'] = self._get_memory_pressure()
-            
-            tprint_success(f"✅ Aggressive memory cleanup completed: {cleanup_results['memory_freed_mb']:.1f}MB freed")
-            
-        except Exception as e:
-            cleanup_results['errors'].append(str(e))
-            self.logger.error(f"❌ Aggressive memory cleanup failed: {e}")
-            tprint_error(f"❌ Aggressive memory cleanup failed: {e}")
-        
-        return cleanup_results
-    
-    def _get_memory_pressure(self) -> float:
-        """Get current memory pressure if available."""
-        if hasattr(self, 'memory_optimizer') and self.memory_optimizer:
-            return getattr(self.memory_optimizer, 'memory_pressure', 0.0)
-        return 0.0
-    
-    def _monitor_memory_pressure(self) -> Dict[str, Any]:
-        """Monitor memory pressure and trigger cleanup if needed."""
-        try:
-            import psutil
-            process = psutil.Process()
-            memory_info = process.memory_info()
-            memory_mb = memory_info.rss / 1024 / 1024
-            
-            # Calculate pressure as ratio of current usage to system memory
-            system_memory = psutil.virtual_memory()
-            pressure = memory_mb / (system_memory.total / 1024 / 1024)
-            
-            memory_stats = {
-                'pressure': pressure,
-                'memory_mb': memory_mb,
-                'cleanup_triggered': False,
-                'recommendations': []
-            }
-            
-            if pressure > self.aggressive_cleanup_threshold:  # Critical pressure
-                memory_stats['cleanup_triggered'] = True
-                memory_stats['recommendations'].append('CRITICAL: Immediate aggressive cleanup required')
-                self._aggressive_memory_cleanup(force_cleanup=True)
-                
-            elif pressure > self.memory_pressure_threshold:  # High pressure
-                memory_stats['cleanup_triggered'] = True
-                memory_stats['recommendations'].append('WARNING: High memory pressure - performing cleanup')
-                self._aggressive_memory_cleanup(force_cleanup=False)
-                
-            elif pressure > 0.7:  # Medium pressure
-                memory_stats['recommendations'].append('INFO: Medium memory pressure - monitoring')
-                self._light_memory_cleanup()
-            
-            return memory_stats
-            
-        except ImportError:
-            # Fallback if psutil not available
-            return {
-                'pressure': 0.0,
-                'memory_mb': 0.0,
-                'cleanup_triggered': False,
-                'recommendations': ['psutil not available - memory monitoring disabled']
-            }
-        except Exception as e:
-            tprint_warning(f"⚠️ Memory monitoring failed: {e}")
-            return {
-                'pressure': 0.0,
-                'memory_mb': 0.0,
-                'cleanup_triggered': False,
-                'recommendations': [f'Memory monitoring error: {e}']
-            }
-
-    def _light_memory_cleanup(self) -> None:
-        """Perform light memory cleanup."""
-        try:
-            # Clear small caches
-            if len(self._cache) > self.max_cache_entries:
-                # Remove oldest 25% of cache entries
-                items_to_remove = len(self._cache) // 4
-                keys_to_remove = list(self._cache.keys())[:items_to_remove]
-                for key in keys_to_remove:
-                    del self._cache[key]
-            
-            # Clear computation cache
-            self._computation_cache.clear()
-            
-            # Force garbage collection
-            import gc
-            gc.collect()
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ Light memory cleanup failed: {e}")
-
-    def _aggressive_memory_cleanup(self, force_cleanup: bool = False) -> Dict[str, Any]:
-        """Perform aggressive memory cleanup."""
-        cleanup_results = {
-            'success': False,
-            'memory_freed_mb': 0.0,
-            'cleanup_methods_used': [],
-            'errors': []
-        }
-        
-        try:
-            # Clear all caches
-            cache_size_before = len(self._cache)
-            self._cache.clear()
-            self._computation_cache.clear()
-            self._variance_cache.clear()
-            cleanup_results['cleanup_methods_used'].append('cache_clear')
-            
-            # Clear vectorized arrays
-            self._vectorized_arrays.clear()
-            cleanup_results['cleanup_methods_used'].append('vectorized_arrays_clear')
-            
-            # Clear component state
-            self._X_cached = None
-            self._last_X = None
-            cleanup_results['cleanup_methods_used'].append('state_clear')
-            
-            # Force garbage collection
-            import gc
-            collected = gc.collect()
-            cleanup_results['cleanup_methods_used'].append('garbage_collection')
-            
-            # Use advanced memory optimizer if available
-            if self.advanced_memory_optimizer:
-                try:
-                    advanced_cleanup = self.advanced_memory_optimizer.aggressive_cleanup(
-                        force_cleanup=force_cleanup,
-                        clear_caches=True,
-                        compress_memory=True
-                    )
-                    cleanup_results['memory_freed_mb'] += advanced_cleanup.get('memory_freed_mb', 0.0)
-                    cleanup_results['cleanup_methods_used'].append('advanced_optimizer')
-                except Exception as e:
-                    cleanup_results['errors'].append(f"Advanced optimizer failed: {e}")
-            
-            cleanup_results['success'] = True
-            cleanup_results['memory_freed_mb'] += cache_size_before * 0.001  # Rough estimate
-            
-        except Exception as e:
-            cleanup_results['errors'].append(str(e))
-            tprint_error(f"❌ Aggressive memory cleanup failed: {e}")
-        
-        return cleanup_results
-
-    def _early_termination_check(self, scores: np.ndarray, threshold: float = 0.001) -> bool:
-        """Check if early termination should be triggered based on score quality."""
-        if len(scores) == 0:
-            return True
-        
-        # Check if maximum score is below threshold
-        max_score = np.max(scores)
-        if max_score < threshold:
-            tprint_warning(f"⚠️ Early termination: max score {max_score:.6f} below threshold {threshold}")
-            return True
-        
-        # Check if too many features have very low scores
-        low_score_count = np.sum(scores < threshold * 0.1)
-        if low_score_count > len(scores) * 0.8:
-            tprint_warning(f"⚠️ Early termination: {low_score_count}/{len(scores)} features have very low scores")
-            return True
-        
-        return False
-
-    def _validate_feature_quality(self, X: pd.DataFrame, y: pd.Series, features: List[str]) -> bool:
-        """Validate that selected features meet quality requirements."""
-        if not features:
-            return False
-        
-        # Check minimum feature count
-        if len(features) < 5:
-            tprint_warning(f"⚠️ Insufficient features selected: {len(features)}")
-            return False
-        
-        # Check feature variance
-        feature_data = X[features]
-        low_variance_features = feature_data.var() < 1e-8
-        if low_variance_features.any():
-            low_var_features = low_variance_features[low_variance_features].index.tolist()
-            tprint_warning(f"⚠️ Low variance features detected: {low_var_features}")
-            return False
-        
-        # Check for constant features
-        constant_features = feature_data.nunique() <= 1
-        if constant_features.any():
-            const_features = constant_features[constant_features].index.tolist()
-            tprint_warning(f"⚠️ Constant features detected: {const_features}")
-            return False
-        
-        return True
-
-    def _trim_cache_by_entries(self, max_entries: int = 1000) -> None:
-        """Trim cache to maximum number of entries (LRU-like)."""
-        if len(self._cache) > max_entries:
-            # Keep only the most recent entries
-            excess = len(self._cache) - max_entries
-            keys_to_remove = list(self._cache.keys())[:excess]
-            
-            for key in keys_to_remove:
-                del self._cache[key]
-            
-            tprint(f"🧹 Trimmed {excess} excess cache entries")
-            self.logger.info(f"Cache trimmed: removed {excess} entries")
-
-    def _set_model_specific_parameters(self):
-        """Set model-specific feature selection parameters."""
-        tprint_debug("🔧 Setting model-specific parameters")
-        model_specific_params = {
-            'AdvancedMambaHybrid': {
-                'model_correlation_threshold': 0.88,  # Allow more correlated features for multi-timeframe fusion
-                'model_importance_threshold': 0.003,  # Moderate threshold for attention mechanisms
-                'min_correlation_threshold': 0.92,   # Tighter initial correlation filtering
-                'min_variance_threshold': 0.02      # Higher variance requirement
-            },
-            'FinancialResNet': {
-                'model_correlation_threshold': 0.95,  # Tighter correlation for regime classification
-                'model_importance_threshold': 0.002,  # Lower threshold for comprehensive input
-                'min_correlation_threshold': 0.96,   # Very tight correlation filtering
-                'min_variance_threshold': 0.01,      # Standard variance requirement
-                'cv_scoring': 'average_precision'  # For regime classification with imbalanced labels
-            },
-            'DeepScaler': {
-                'model_correlation_threshold': 0.85,  # Looser correlation for precision focus
-                'model_importance_threshold': 0.008,  # Higher threshold for cleaner features
-                'min_correlation_threshold': 0.98,   # Very tight correlation filtering
-                'min_variance_threshold': 0.03      # Higher variance requirement
-            },
-            'NBEATS': {
-                'model_correlation_threshold': 0.90,  # Moderate correlation for time series
-                'model_importance_threshold': 0.005,  # Standard threshold for temporal modeling
-                'min_correlation_threshold': 0.94,   # Tight correlation for clean time series
-                'min_variance_threshold': 0.015     # Moderate variance requirement
-            }
-        }
-
-        if self.config.model_type in model_specific_params:
-            params = model_specific_params[self.config.model_type]
-            for param, value in params.items():
-                setattr(self.config, param, value)
-            tprint(f"✅ Applied {self.config.model_type} specific parameters")
-
-    def _get_cache_key(self, operation: str, data_hash: str, params: Dict[str, Any] = None) -> str:
-        """Generate a cache key for an operation."""
-        params_str = str(sorted(params.items())) if params else ""
-        return f"{operation}_{data_hash}_{params_str}"
-
-    def _get_data_hash(self, X: pd.DataFrame, y: pd.Series = None) -> str:
-        """Generate a hash for the input data."""
-        data_str = f"{X.shape}_{X.columns.tolist()}"
-        if y is not None:
-            data_str += f"_{y.shape}_{y.dtype}"
-        return hashlib.md5(data_str.encode()).hexdigest()[:16]
-
-    def _vectorized_correlation_analysis(self, X: pd.DataFrame) -> np.ndarray:
-        """Vectorized correlation analysis using VectorBT or numpy operations."""
-        cache_key = self._get_cache_key("correlation", self._get_data_hash(X))
-        
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            tprint(f"📊 Cache hit for correlation analysis (hit rate: {self._cache_hits/(self._cache_hits+self._cache_misses):.2%})")
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        tprint("🔄 Computing vectorized correlation matrix...")
-        
-        try:
-            X_numeric = X.select_dtypes(include=[np.number])
-            
-            if VECTORBT_AVAILABLE and X_numeric.shape[1] > 50:  # Use VectorBT for large feature sets
-                tprint("🚀 Using VectorBT for correlation analysis")
-                # Use VectorBT's optimized correlation calculation
-                corr_matrix = vbt.correlation_matrix(X_numeric.values, method='pearson')
-                corr_matrix = corr_matrix.values
-            else:
-                # Use numpy for smaller datasets or when VectorBT is not available
-                tprint("📊 Using numpy for correlation analysis")
-                corr_matrix = np.corrcoef(X_numeric.T)
-            
-            # Handle NaN values
-            corr_matrix = np.nan_to_num(corr_matrix, nan=0.0)
-            
-            # Cache the result
-            self._cache[cache_key] = corr_matrix
-            tprint(f"✅ Vectorized correlation computed: {corr_matrix.shape}")
-            
-            return corr_matrix
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ Correlation analysis failed: {e}")
-            # Return identity matrix as fallback
-            return np.eye(X_numeric.shape[1] if 'X_numeric' in locals() else X.shape[1])
-
-    def _vectorized_variance_analysis(self, X: pd.DataFrame) -> np.ndarray:
-        """Vectorized variance analysis using numpy operations."""
-        cache_key = self._get_cache_key("variance", self._get_data_hash(X))
-        
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        tprint("🔄 Computing vectorized variance analysis...")
-        
-        # Use numpy for vectorized variance computation
-        X_numeric = X.select_dtypes(include=[np.number])
-        variances = np.var(X_numeric.values, axis=0, ddof=1)
-        
-        # Cache the result
-        self._cache[cache_key] = variances
-        tprint(f"✅ Vectorized variance computed: {len(variances)} features")
-        
-        return variances
-
-    def _vectorized_feature_importance(self, X: pd.DataFrame, y: pd.Series, model_type: str = 'rf') -> np.ndarray:
-        """Vectorized feature importance computation with VectorBT optimization."""
-        cache_key = self._get_cache_key("importance", self._get_data_hash(X, y), {"model_type": model_type})
-        
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        tprint(f"🔄 Computing vectorized feature importance using {model_type}...")
-        
-        try:
-            if VECTORBT_AVAILABLE and X.shape[1] > 100:  # Use VectorBT for large feature sets
-                tprint("🚀 Using VectorBT for feature importance calculation")
-                # Use VectorBT's optimized operations for feature importance
-                importance = self._vectorbt_feature_importance(X, y, model_type)
-            else:
-                # Use traditional method for smaller datasets
-                tprint("📊 Using traditional method for feature importance")
-                importance = self._traditional_feature_importance(X, y, model_type)
-            
-            # Cache the result
-            self._cache[cache_key] = importance
-            tprint(f"✅ Vectorized importance computed: {len(importance)} features")
-            
-            return importance
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ Feature importance calculation failed: {e}")
-            # Fallback to traditional method
-            return self._traditional_feature_importance(X, y, model_type)
-    
-    def _vectorbt_feature_importance(self, X: pd.DataFrame, y: pd.Series, model_type: str = 'rf') -> np.ndarray:
-        """VectorBT-optimized feature importance calculation."""
-        try:
-            # Use VectorBT's vectorized operations for faster computation
-            if model_type == 'rf':
-                # Use VectorBT's batch processing for RandomForest
-                importance = vbt.rolling_apply(
-                    X.values, 
-                    lambda x: self._train_random_forest(pd.DataFrame(x, columns=X.columns), y).feature_importances_,
-                    window=min(1000, len(X)),  # Process in chunks
-                    step=1
-                )
-                # Take the mean across all windows
-                importance = np.mean(importance, axis=0)
-            else:
-                # Fallback to traditional method for other models
-                importance = self._traditional_feature_importance(X, y, model_type)
-            
-            return importance
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ VectorBT feature importance failed: {e}")
-            return self._traditional_feature_importance(X, y, model_type)
-    
-    def _traditional_feature_importance(self, X: pd.DataFrame, y: pd.Series, model_type: str = 'rf') -> np.ndarray:
-        """Traditional feature importance calculation."""
-        # Train model and get importance
-        if model_type == 'rf':
-            model = self._train_random_forest(X, y)
-        elif model_type == 'lightgbm' and LIGHTGBM_AVAILABLE:
-            model = self._train_lightgbm_model(X, y)
-        else:
-            model = self._train_random_forest(X, y)
-        
-        if hasattr(model, 'feature_importances_'):
-            importance = model.feature_importances_
-        else:
-            raise AttributeError(f"Model {type(model).__name__} does not have feature_importances_ attribute. Cannot compute feature importance.")
-        
-        return importance
-
-    def _vectorized_mutual_information(self, X: pd.DataFrame, y: pd.Series) -> np.ndarray:
-        """Vectorized mutual information computation."""
-        
-        cache_key = self._get_cache_key("mutual_info", self._get_data_hash(X, y))
-        
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        tprint("🔄 Computing vectorized mutual information...")
-        
-        try:
-            from sklearn.feature_selection import mutual_info_regression, mutual_info_classif
-            
-            if self._is_classification(y):
-                mi_scores = mutual_info_classif(X, y, random_state=42)
-            else:
-                mi_scores = mutual_info_regression(X, y, random_state=42)
-            
-            # Normalize scores
-            max_mi = np.max(mi_scores) if len(mi_scores) > 0 else 1.0
-            normalized_scores = mi_scores / max_mi if max_mi > 0 else mi_scores
-            
-            # Cache the result
-            self._cache[cache_key] = normalized_scores
-            tprint(f"✅ Vectorized mutual information computed: {len(normalized_scores)} features")
-            
-            return normalized_scores
-            
-        except ImportError as e:
-            raise ImportError("sklearn is required for mutual information calculation but not available. Please install sklearn") from e
-        except Exception as e:
-            raise RuntimeError(f"Mutual information calculation failed: {e}") from e
-
-    def _vectorized_stability_analysis(self, X: pd.DataFrame, y: pd.Series) -> np.ndarray:
-        """Vectorized stability analysis across time periods."""
-        if len(X) < 100:
-            return np.ones(len(X.columns))  # Default stability for small datasets
-        
-        cache_key = self._get_cache_key("stability", self._get_data_hash(X, y))
-        
-        if cache_key in self._cache:
-            self._cache_hits += 1
-            return self._cache[cache_key]
-        
-        self._cache_misses += 1
-        tprint("🔄 Computing vectorized stability analysis...")
-        
-        try:
-            # Split data into chunks for stability analysis
-            chunk_size = min(500, len(X) // 3)
-            if chunk_size < 50:
-                return np.ones(len(X.columns))
-            
-            # Create overlapping chunks
-            chunk_indices = []
-            for i in range(0, len(X) - chunk_size + 1, chunk_size // 2):
-                chunk_indices.append((i, i + chunk_size))
-            
-            if len(chunk_indices) < 2:
-                return np.ones(len(X.columns))
-            
-            # Calculate importance for each chunk using vectorized operations
-            chunk_importances = []
-            for start_idx, end_idx in chunk_indices:
-                try:
-                    X_chunk = X.iloc[start_idx:end_idx]
-                    y_chunk = y.iloc[start_idx:end_idx]
-                    
-                    # Use vectorized importance calculation
-                    importance = self._vectorized_feature_importance(X_chunk, y_chunk)
-                    chunk_importances.append(importance)
-                except Exception:
-                    continue
-            
-            if len(chunk_importances) < 2:
-                return np.ones(len(X.columns))
-            
-            # Calculate stability as consistency across chunks
-            chunk_importances = np.array(chunk_importances)
-            
-            # Normalize each chunk's importance
-            normalized_chunks = chunk_importances / (np.max(chunk_importances, axis=1, keepdims=True) + 1e-8)
-            
-            # Calculate stability as 1 - coefficient of variation
-            mean_importance = np.mean(normalized_chunks, axis=0)
-            std_importance = np.std(normalized_chunks, axis=0)
-            stability_scores = 1.0 / (1.0 + std_importance / (mean_importance + 1e-8))
-            
-            # Cache the result
-            self._cache[cache_key] = stability_scores
-            tprint(f"✅ Vectorized stability analysis computed: {len(stability_scores)} features")
-            
-            return stability_scores
-            
-        except Exception as e:
-            tprint(f"⚠️ Stability analysis failed: {e}")
-            return np.ones(len(X.columns))
-
-    def _clear_cache(self):
-        """Clear the computation cache with thread safety."""
-        tprint_debug("🧹 Clearing computation cache")
-        with self._cache_lock:
-            cache_size = len(self._cache)
-            self._cache.clear()
-            self._computation_cache.clear()
-            self._cache_hits = 0
-            self._cache_misses = 0
-            tprint(f"🧹 Cache cleared: {cache_size} entries removed")
-
-    def _get_from_cache(self, key: str) -> Any:
-        """Thread-safe cache retrieval."""
-        with self._cache_lock:
-            if key in self._cache:
-                self._cache_hits += 1
-                return self._cache[key]
-            else:
-                self._cache_misses += 1
-                return None
-
-    def _set_cache(self, key: str, value: Any) -> None:
-        """Thread-safe cache storage with size limits."""
-        with self._cache_lock:
-            # Check cache size limits
-            if len(self._cache) >= self.max_cache_entries:
-                # Remove oldest 25% of entries
-                items_to_remove = len(self._cache) // 4
-                keys_to_remove = list(self._cache.keys())[:items_to_remove]
-                for k in keys_to_remove:
-                    del self._cache[k]
-            
-            self._cache[key] = value
-
-    def _get_cache_stats(self) -> Dict[str, Any]:
-        """Get cache statistics."""
-        total_requests = self._cache_hits + self._cache_misses
-        hit_rate = self._cache_hits / total_requests if total_requests > 0 else 0.0
-
-        stats = {
-            'cache_size': len(self._cache),
-            'cache_hits': self._cache_hits,
-            'cache_misses': self._cache_misses,
-            'hit_rate': hit_rate,
-            'computation_cache_size': len(self._computation_cache)
-        }
-
-        tprint(
-            (
-                "💾 Cache stats — size: {cache_size}, hits: {cache_hits}, misses: {cache_misses}, hit rate: {hit_rate:.2%},"
-                " computation cache: {computation_cache_size}"
-            ).format(**stats)
-        )
-
-        return stats
-    
-    def select_features(self,
-                       X: pd.DataFrame,
-                       y: pd.Series,
-                       feature_names: Optional[List[str]] = None,
-                       event_timestamps: Optional[Union[pd.Series, pd.Index, List[pd.Timestamp], np.ndarray]] = None) -> FeatureSelectionResult:
-        """Perform multi-stage feature selection with directional support and vectorization."""
-
-        start_time = time.time()
-        tprint("🔍 Starting multi-stage feature selection with vectorization and caching")
-        
-        # Display VectorBT optimization status
-        self._display_vectorbt_status()
-        
-        try:
-            # Fast fail on input validation
-            self._validate_inputs(X, y)
-            
-            tprint(f"📊 Input data: {X.shape[0]} samples, {X.shape[1]} features")
-            tprint(f"🎯 Target variable: {y.shape[0]} samples, type: {'classification' if self._is_classification(y) else 'regression'}")
-
-            # Apply VectorBT memory optimization if available
-            if VECTORBT_AVAILABLE and X.shape[1] > 100:
-                tprint("🧠 Applying VectorBT memory optimization")
-                X = self._vectorbt_memory_optimization(X, "feature_selection_input")
-
-            # Clear cache at start of new selection
-            self._clear_cache()
-            self._active_event_timestamps = None
-        
-            # Monitor memory pressure at start
-            memory_stats = self._monitor_memory_pressure()
-            if memory_stats['cleanup_triggered']:
-                tprint(f"🧠 Memory pressure detected at start: {memory_stats['pressure']:.2f}")
-                for recommendation in memory_stats['recommendations']:
-                    tprint(f"   → {recommendation}")
-
-            # Align event timestamps to the provided data (before any windowing)
-            event_ts_series_full: Optional[pd.Series] = None
-            if event_timestamps is not None:
-                tprint("🕐 Processing event timestamps for temporal alignment")
-                if isinstance(event_timestamps, pd.Series):
-                    aligned = event_timestamps
-                    if not aligned.index.equals(X.index):
-                        aligned = aligned.reindex(X.index)
-                    if aligned.isna().any():
-                        raise ValueError("Event timestamps are missing for some samples after alignment.")
-                    event_ts_series_full = pd.to_datetime(aligned)
-                    tprint(f"   ✅ Aligned {len(aligned)} event timestamps")
-                else:
-                    event_index = pd.Index(event_timestamps)
-                    if len(event_index) != len(X.index):
-                        raise ValueError("Event timestamps length must match the number of samples.")
-                    event_ts_series_full = pd.Series(pd.to_datetime(event_index), index=X.index)
-                    tprint(f"   ✅ Created event timestamp series with {len(event_index)} timestamps")
-            elif isinstance(X.index, pd.DatetimeIndex):
-                event_ts_series_full = pd.Series(X.index, index=X.index)
-                tprint("   ✅ Using DataFrame index as event timestamps")
-
-        # Apply execution mode data windowing if configured
-        X_processed = X.copy()
-        if self.execution_mode_config:
-            window_days = self.execution_mode_config.get('window_days', 1460)
-            if len(X_processed) > window_days:
-                X_processed = X_processed.tail(window_days).copy()
-                tprint(f"📊 Applied execution mode window: using last {window_days} samples for feature selection")
-
-        if event_ts_series_full is not None:
-            event_ts_series = event_ts_series_full.reindex(X_processed.index)
-            if event_ts_series.isna().any():
-                raise ValueError("Event timestamps missing after applying execution mode window.")
-            self._active_event_timestamps = pd.to_datetime(event_ts_series)
-        elif isinstance(X_processed.index, pd.DatetimeIndex):
-            self._active_event_timestamps = pd.Series(X_processed.index, index=X_processed.index)
-
-        # Apply execution mode stage targets if configured
-        default_stage_targets = (
-            self.config.initial_features,
-            self.config.stage_1_target,
-            self.config.stage_2_target,
-            self.config.stage_3_target,
-        )
-
-        stage_targets: Tuple[int, ...] = default_stage_targets
-        if self.execution_mode_config:
-            raw_stage_targets = self.execution_mode_config.get('stage_targets', default_stage_targets)
-
-            if isinstance(raw_stage_targets, (list, tuple)):
-                padded_targets = list(default_stage_targets)
-                for idx, value in enumerate(raw_stage_targets):
-                    if idx < len(padded_targets):
-                        padded_targets[idx] = value
-                    else:
-                        padded_targets.append(value)
-                stage_targets = tuple(padded_targets)
-            elif isinstance(raw_stage_targets, int):
-                padded_targets = list(default_stage_targets)
-                padded_targets[-1] = raw_stage_targets
-                stage_targets = tuple(padded_targets)
-            else:
-                stage_targets = default_stage_targets
-
-            tprint(f"📊 Using execution mode stage targets: {stage_targets}")
-
-        # Log the feature reduction pipeline
-        tprint("🎯 FEATURE REDUCTION PIPELINE:")
-        tprint(f"   📊 Initial Features: {stage_targets[0]}")
-        tprint(f"   📊 Stage 1 Target: {stage_targets[1]} features")
-        tprint(f"   📊 Stage 2 Target: {stage_targets[2]} features")
-        tprint(f"   📊 Stage 3 Target: {stage_targets[3]} features")
-        tprint(f"   📊 Final Target: {stage_targets[3]} features")
-
-        # Determine which directions to process based on mode
-        directions_to_process = self._get_directions_to_process()
-
-        # Handle separate directional feature selection
-        if self.config.separate_directional_features and self.config.direction_mode in ['both', 'long_only', 'short_only']:
-            tprint(f"🎯 Processing directional features: {directions_to_process}")
-            result = self._select_directional_features(X_processed, y, feature_names, directions_to_process, stage_targets)
-        else:
-            tprint("🎯 Processing unified feature selection")
-            result = self._select_unified_features(X_processed, y, feature_names, stage_targets)
-
-        # Log final cache statistics
-        cache_stats = self._get_cache_stats()
-        tprint("📊 CACHE STATISTICS:")
-        tprint(f"   💾 Cache size: {cache_stats['cache_size']} entries")
-        tprint(f"   ✅ Cache hits: {cache_stats['cache_hits']}")
-        tprint(f"   ❌ Cache misses: {cache_stats['cache_misses']}")
-        tprint(f"   📈 Hit rate: {cache_stats['hit_rate']:.2%}")
-
-        # Perform final memory cleanup
-        final_cleanup = self._aggressive_memory_cleanup(force_cleanup=False)
-        tprint("🧹 FINAL MEMORY CLEANUP:")
-        tprint(f"   💾 Memory freed: {final_cleanup['memory_freed_mb']:.1f}MB")
-        tprint(f"   ✅ Cleanup success: {final_cleanup['success']}")
-        if final_cleanup['cleanup_methods_used']:
-            tprint(f"   🔧 Methods used: {', '.join(final_cleanup['cleanup_methods_used'])}")
-
-        # Log execution time
-        execution_time = time.time() - start_time
-        result.selection_time = execution_time
-        tprint(f"⏱️ Total execution time: {execution_time:.3f} seconds")
-
-        # Ensure polarity adjustments are attached to the result payload
-        result.polarity_adjustments = getattr(self, 'feature_polarity_adjustments', {})
-        result.sign_stability = getattr(self, 'feature_sign_stability', {})
-
-        return result
-
-        except ValueError as ve:
-            # Fast fail on validation errors
-            tprint_error(f"❌ Input validation failed: {ve}")
-            raise ve
-        except MemoryError as me:
-            # Handle memory errors with cleanup
-            tprint_error(f"❌ Memory error during feature selection: {me}")
-            self._aggressive_memory_cleanup(force_cleanup=True)
-            raise RuntimeError(f"Feature selection failed due to memory constraints: {me}") from me
-        except Exception as e:
-            # Handle unexpected errors with proper logging
-            tprint_error(f"❌ Unexpected error during feature selection: {e}")
-            import traceback
-            tprint_error(f"🔍 Error details: {traceback.format_exc()}")
-            
-            # Perform cleanup on error
-            try:
-                self._aggressive_memory_cleanup(force_cleanup=True)
-            except Exception as cleanup_error:
-                tprint_warning(f"⚠️ Cleanup failed after error: {cleanup_error}")
-            
-            # Return empty result with error information
-            error_result = FeatureSelectionResult()
-            error_result.eligible_for_selection = False
-            error_result.turnover_rejection_reason = f"Selection failed: {str(e)}"
-            return error_result
-
-    def _get_directions_to_process(self) -> List[str]:
-        """Determine which directions to process based on configuration."""
-        if self.config.direction_mode == 'long_only':
-            return ['long']
-        elif self.config.direction_mode == 'short_only':
-            return ['short']
-        else:  # 'both' or other
-            return ['long', 'short']
-
-    def _get_event_timestamps_for_index(self, index: pd.Index) -> Optional[pd.Series]:
-        """Return aligned event timestamps for the provided index."""
-
-        event_ts = getattr(self, '_active_event_timestamps', None)
-        if event_ts is not None:
-            aligned = event_ts.reindex(index)
-            if aligned.isna().any():
-                self.logger.warning(
-                    "⚠️ Missing event timestamps after alignment; falling back to DataFrame index ordering."
-                )
-                if isinstance(index, pd.DatetimeIndex):
-                    return pd.Series(index, index=index)
-                return None
-            return aligned
-
-        if isinstance(index, pd.DatetimeIndex):
-            return pd.Series(index, index=index)
-
-        return None
-
-    def _select_directional_features(self,
-                                   X: pd.DataFrame,
-                                   y: pd.Series,
-                                   feature_names: Optional[List[str]],
-                                   directions: List[str],
-                                   stage_targets: Tuple[int, ...]) -> FeatureSelectionResult:
-        """Select features separately for each direction."""
-
-        self.logger.info(f"🎯 Selecting directional features for: {directions}")
-        tprint(f"🎯 Directional selection activated for: {directions}")
-
-        # Process each direction separately
-        for direction in directions:
-            self.logger.info(f"🔄 Processing {direction} direction features")
-            tprint(f"🔄 Processing direction: {direction}")
-
-            # Filter features for this direction
-            tprint(f"   🔍 Filtering features for {direction} direction")
-            direction_features = self._filter_direction_features(X, direction)
-            tprint(f"   📊 Found {len(direction_features.columns)} {direction} features")
-
-            if len(direction_features.columns) < stage_targets[-1]:
-                tprint_error(f"❌ Insufficient features for {direction} direction: {len(direction_features.columns)} available, requires at least {stage_targets[-1]}")
-                raise ValueError(
-                    f"Insufficient features for {direction} direction: {len(direction_features.columns)} available, "
-                    f"requires at least {stage_targets[-1]}"
-                )
-
-            # Select features for this direction
-            tprint(f"   🎯 Running unified feature selection for {direction} direction")
-            direction_result = self._select_unified_features(direction_features, y, feature_names, stage_targets)
-            tprint(f"   ✅ {direction} direction selection completed: {len(direction_result.final_features)} final features")
-
-            # Store results for this direction
-            if direction == 'long':
-                self.long_results = direction_result
-                tprint(f"   💾 Stored long direction results")
-            else:
-                self.short_results = direction_result
-                tprint(f"   💾 Stored short direction results")
-
-        # Compile combined results
-        return self._compile_directional_results(directions)
-
-    def _filter_direction_features(self, X: pd.DataFrame, direction: str) -> pd.DataFrame:
-        """Filter features for a specific direction."""
-        prefix = self.config.directional_feature_prefixes.get(direction, f'{direction}_')
-
-        # Find features that start with the direction prefix
-        direction_features = X.columns[X.columns.str.startswith(prefix)]
-
-        if len(direction_features) == 0:
-            self.logger.warning(f"⚠️ No {direction} features found with prefix '{prefix}'")
-            # Fallback: use all features if no prefixed features found
-            return X.copy()
-
-        self.logger.info(f"📊 Found {len(direction_features)} {direction} features")
-        return X[direction_features]
-
-    def _compile_directional_results(self, directions: List[str]) -> FeatureSelectionResult:
-        """Compile results from directional feature selection."""
-
-        combined_result = FeatureSelectionResult()
-        tprint("📦 Compiling directional selection results")
-
-        if 'long' in directions and hasattr(self, 'long_results'):
-            # Combine long and short features
-            long_features = self.long_results.final_features
-            short_features = self.short_results.final_features if 'short' in directions else []
-
-            # Ensure no overlap between long and short features
-            combined_features = long_features + short_features
-
-            combined_result.final_features = combined_features
-            combined_result.feature_counts = {
-                'long_features': len(long_features),
-                'short_features': len(short_features),
-                'total_features': len(combined_features),
-                'long_stage_1': len(self.long_results.stage_1_features),
-                'long_stage_2': len(self.long_results.stage_2_features),
-                'long_stage_3': len(self.long_results.stage_3_features),
-                'short_stage_1': len(self.short_results.stage_1_features) if 'short' in directions else 0,
-                'short_stage_2': len(self.short_results.stage_2_features) if 'short' in directions else 0,
-                'short_stage_3': len(self.short_results.stage_3_features) if 'short' in directions else 0,
-            }
-
-        combined_result.selection_time = time.time() - time.time()  # Will be updated by caller
-
-        self.logger.info(f"✅ Directional feature selection completed")
-        self.logger.info(f"📊 Long features: {len(long_features) if 'long' in directions else 0}")
-        self.logger.info(f"📊 Short features: {len(short_features) if 'short' in directions else 0}")
-        self.logger.info(f"📊 Total unique features: {len(combined_features)}")
-        directional_summary = (
-            f"✅ Directional selection complete — long: {len(long_features) if 'long' in directions else 0},"
-            f" short: {len(short_features) if 'short' in directions else 0}, total: {len(combined_features)}"
-        )
-        tprint(directional_summary)
-
-        # Generate outcome file with datetime stamp (directional)
-        try:
-            from datetime import datetime
-            from pathlib import Path
-            import json
-            
-            outcome_timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-            outcomes_dir = Path('outcomes')
-            outcomes_dir.mkdir(parents=True, exist_ok=True)
-            
-            outcome_filename = f"final_feature_selection_directional_outcome_{outcome_timestamp}.json"
-            outcome_path = outcomes_dir / outcome_filename
-            
-            # Feature overlap analysis
-            feature_overlap = {}
-            if 'long' in directions and 'short' in directions:
-                long_set = set(long_features)
-                short_set = set(short_features)
-                overlap = long_set & short_set
-                feature_overlap = {
-                    'overlap_count': len(overlap),
-                    'overlap_features': list(overlap),
-                    'unique_to_long': len(long_set - short_set),
-                    'unique_to_short': len(short_set - long_set),
-                    'overlap_percentage': float(len(overlap) / max(1, len(long_set | short_set)) * 100),
-                }
-            
-            # Stage transition analysis
-            stage_transitions = {
-                'long_pipeline': {
-                    'stage_1_to_2_reduction': self.long_results.feature_counts.get('long_stage_1', 0) - self.long_results.feature_counts.get('long_stage_2', 0) if 'long' in directions else 0,
-                    'stage_2_to_3_reduction': self.long_results.feature_counts.get('long_stage_2', 0) - self.long_results.feature_counts.get('long_stage_3', 0) if 'long' in directions else 0,
-                    'total_reduction_pct': float((1 - self.long_results.feature_counts.get('long_stage_3', 0) / max(1, self.long_results.feature_counts.get('long_stage_1', 1))) * 100) if 'long' in directions else 0,
-                },
-                'short_pipeline': {
-                    'stage_1_to_2_reduction': self.short_results.feature_counts.get('short_stage_1', 0) - self.short_results.feature_counts.get('short_stage_2', 0) if 'short' in directions else 0,
-                    'stage_2_to_3_reduction': self.short_results.feature_counts.get('short_stage_2', 0) - self.short_results.feature_counts.get('short_stage_3', 0) if 'short' in directions else 0,
-                    'total_reduction_pct': float((1 - self.short_results.feature_counts.get('short_stage_3', 0) / max(1, self.short_results.feature_counts.get('short_stage_1', 1))) * 100) if 'short' in directions else 0,
-                }
-            }
-            
-            # Create comprehensive outcome report
-            outcome_data = {
-                'component': 'final_feature_selection_directional',
-                'timestamp': datetime.now().isoformat(),
-                'execution_time': combined_result.selection_time,
-                'selection_type': 'directional',
-                'directions': directions,
-                'configuration': {
-                    'enable_directional_features': self.config.enable_directional_features,
-                    'stage_1_target': self.config.stage_1_target,
-                    'stage_2_target': self.config.stage_2_target,
-                    'stage_3_target': self.config.stage_3_target,
-                    'n_estimators': self.config.n_estimators,
-                    'cv_folds': self.config.cv_folds,
-                },
-                'results': {
-                    'summary': {
-                        'long_features_count': len(long_features) if 'long' in directions else 0,
-                        'short_features_count': len(short_features) if 'short' in directions else 0,
-                        'total_features_count': len(combined_features),
-                    },
-                    'long_features': long_features if 'long' in directions else [],
-                    'short_features': short_features if 'short' in directions else [],
-                    'feature_counts': combined_result.feature_counts,
-                    'feature_overlap': feature_overlap,
-                    'stage_transitions': stage_transitions,
-                },
-                'long_pipeline_scores': {
-                    'stage_1': self.long_results.stage_1_scores if 'long' in directions else {},
-                    'stage_2': self.long_results.stage_2_scores if 'long' in directions else {},
-                    'stage_3': self.long_results.stage_3_scores if 'long' in directions else {},
-                },
-                'short_pipeline_scores': {
-                    'stage_1': self.short_results.stage_1_scores if 'short' in directions else {},
-                    'stage_2': self.short_results.stage_2_scores if 'short' in directions else {},
-                    'stage_3': self.short_results.stage_3_scores if 'short' in directions else {},
-                },
-                'status': 'success'
-            }
-            
-            # Save outcome file
-            with open(outcome_path, 'w') as f:
-                json.dump(outcome_data, f, indent=2, default=str)
-            
-            tprint(f"📄 Outcome file saved: {outcome_filename}")
-            
-        except Exception as outcome_error:
-            tprint(f"⚠️ Failed to save outcome file: {outcome_error}")
-            # Don't fail the component if outcome file generation fails
-
-        return combined_result
-
-    def _select_unified_features(self, X: pd.DataFrame, y: pd.Series, feature_names: Optional[List[str]], stage_targets: Tuple[int, ...]) -> FeatureSelectionResult:
-        """Original unified feature selection logic."""
-
-        # Validate inputs
-        tprint("🔍 Validating inputs for unified feature selection")
-        if len(X.columns) < stage_targets[-1]:  # Check against final target
-            tprint_error(f"❌ Insufficient features for unified selection: {len(X.columns)} available, requires at least {stage_targets[-1]}")
-            raise ValueError(
-                f"Insufficient features for unified selection: {len(X.columns)} available, "
-                f"requires at least {stage_targets[-1]}"
-            )
-        tprint(f"✅ Input validation passed: {len(X.columns)} features available")
-
-        # Stage 0: Initial feature preparation
-        self.logger.info("📊 Stage 0: Initial feature preparation")
-        tprint("📊 Stage 0 — preparing initial features")
-        tprint(f"   📊 Input shape: {X.shape[0]} samples, {X.shape[1]} features")
-        prepared_features = self._prepare_initial_features(X, y, feature_names)
-        tprint(f"   ✅ Prepared features: {prepared_features.shape[0]} samples, {prepared_features.shape[1]} features")
-
-        # Stage 1: Initial → Stage 1 target features (Vectorized mRMR + Spearman)
-        stage_1_target = stage_targets[1] if len(stage_targets) > 1 else 100
-        self.logger.info(f"📊 Stage 1: Reducing to {stage_1_target} features")
-        tprint(f"🚀 Stage 1 target: {stage_1_target} features")
-        tprint(f"   📊 Input: {prepared_features.shape[1]} features → Target: {stage_1_target} features")
-        stage_1_features, stage_1_scores = self._stage_1_selection(prepared_features, y, target_count=stage_1_target)
-        tprint(f"   ✅ Stage 1 completed: {len(stage_1_features)} features selected")
-
-        # Stage 2: Stage 1 → Stage 2 target features (Vectorized iterative bottom removal)
-        stage_2_target = stage_targets[2] if len(stage_targets) > 2 else 80
-        self.logger.info(f"📊 Stage 2: Reducing to {stage_2_target} features")
-        tprint(f"🚀 Stage 2 target: {stage_2_target} features")
-        tprint(f"   📊 Input: {len(stage_1_features)} features → Target: {stage_2_target} features")
-        stage_2_features, stage_2_scores = self._stage_2_selection(prepared_features[stage_1_features], y, target_count=stage_2_target)
-        tprint(f"   ✅ Stage 2 completed: {len(stage_2_features)} features selected")
-
-        # Stage 3: Stage 2 → Stage 3 target features (Vectorized chunked RFE + Stability)
-        stage_3_target = stage_targets[3] if len(stage_targets) > 3 else 60
-        self.logger.info(f"📊 Stage 3: Reducing to {stage_3_target} features")
-        tprint(f"🚀 Stage 3 target: {stage_3_target} features")
-        tprint(f"   📊 Input: {len(stage_2_features)} features → Target: {stage_3_target} features")
-        stage_3_features, stage_3_scores = self._stage_3_selection(prepared_features[stage_1_features][stage_2_features], y, target_count=stage_3_target)
-        tprint(f"   ✅ Stage 3 completed: {len(stage_3_features)} features selected")
-
-        # Compile final results
-        tprint("📦 Compiling multi-stage selection results")
-        self._compile_results(
-            prepared_features, y, stage_1_features, stage_2_features, stage_3_features,
-            stage_1_scores, stage_2_scores, stage_3_scores
-        )
-
-        # Save results
-        if self.config.save_analysis:
-            tprint("💾 Saving feature selection analysis")
-            self._save_analysis()
-
-        # Generate outcome file with datetime stamp (unified)
-        try:
-            from datetime import datetime
-            from pathlib import Path
-            import json
-            
-            outcome_timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-            outcomes_dir = Path('outcomes')
-            outcomes_dir.mkdir(parents=True, exist_ok=True)
-            
-            outcome_filename = f"final_feature_selection_outcome_{outcome_timestamp}.json"
-            outcome_path = outcomes_dir / outcome_filename
-            
-            # Stage transition analysis
-            stage_transitions = {
-                'stage_1_to_2': {
-                    'features_removed': len(self.results.stage_1_features) - len(self.results.stage_2_features),
-                    'removal_rate_pct': float((len(self.results.stage_1_features) - len(self.results.stage_2_features)) / max(1, len(self.results.stage_1_features)) * 100),
-                },
-                'stage_2_to_3': {
-                    'features_removed': len(self.results.stage_2_features) - len(self.results.stage_3_features),
-                    'removal_rate_pct': float((len(self.results.stage_2_features) - len(self.results.stage_3_features)) / max(1, len(self.results.stage_2_features)) * 100),
-                },
-                'overall': {
-                    'total_features_removed': len(self.results.stage_1_features) - len(self.results.final_features),
-                    'total_removal_rate_pct': float((len(self.results.stage_1_features) - len(self.results.final_features)) / max(1, len(self.results.stage_1_features)) * 100),
-                }
-            }
-            
-            # Feature importance summary (top features from each stage)
-            feature_importance_summary = {}
-            if self.results.stage_1_scores and 'feature_importance' in self.results.stage_1_scores:
-                importance_dict = self.results.stage_1_scores.get('feature_importance', {})
-                if isinstance(importance_dict, dict):
-                    sorted_features = sorted(importance_dict.items(), key=lambda x: x[1], reverse=True)
-                    feature_importance_summary['top_10_stage_1'] = [
-                        {'feature': feat, 'importance': float(imp)} 
-                        for feat, imp in sorted_features[:10]
-                    ]
-            
-            # Create comprehensive outcome report
-            outcome_data = {
-                'component': 'final_feature_selection',
-                'timestamp': datetime.now().isoformat(),
-                'execution_time': self.results.selection_time,
-                'selection_type': 'unified',
-                'configuration': {
-                    'stage_1_target': self.config.stage_1_target,
-                    'stage_2_target': self.config.stage_2_target,
-                    'stage_3_target': self.config.stage_3_target,
-                    'n_estimators': self.config.n_estimators,
-                    'cv_folds': self.config.cv_folds,
-                    'enable_caching': self.config.enable_caching,
-                    'cache_size': self.config.cache_size,
-                },
-                'results': {
-                    'summary': {
-                        'stage_1_features_count': len(self.results.stage_1_features),
-                        'stage_2_features_count': len(self.results.stage_2_features),
-                        'stage_3_features_count': len(self.results.stage_3_features),
-                        'final_features_count': len(self.results.final_features),
-                    },
-                    'stage_1_features': self.results.stage_1_features,
-                    'stage_2_features': self.results.stage_2_features,
-                    'stage_3_features': self.results.stage_3_features,
-                    'final_features': self.results.final_features,
-                    'feature_counts': self.results.feature_counts,
-                    'stage_transitions': stage_transitions,
-                    'feature_importance_summary': feature_importance_summary,
-                },
-                'quality_metrics': {
-                    'stage_1_scores': self.results.stage_1_scores,
-                    'stage_2_scores': self.results.stage_2_scores,
-                    'stage_3_scores': self.results.stage_3_scores,
-                },
-                'polarity_adjustments': self.results.polarity_adjustments if hasattr(self.results, 'polarity_adjustments') else {},
-                'sign_stability': self.results.sign_stability if hasattr(self.results, 'sign_stability') else {},
-                'status': 'success'
-            }
-            
-            # Save outcome file
-            with open(outcome_path, 'w') as f:
-                json.dump(outcome_data, f, indent=2, default=str)
-            
-            tprint(f"📄 Outcome file saved: {outcome_filename}")
-            
-        except Exception as outcome_error:
-            tprint(f"⚠️ Failed to save outcome file: {outcome_error}")
-            # Don't fail the component if outcome file generation fails
-
-        return self.results
-
-    def _evaluate_feature_polarity(self, X: pd.DataFrame, y: pd.Series) -> Tuple[Dict[str, Dict[str, Any]], Dict[str, List[float]], Dict[str, float]]:
-        """Evaluate rolling feature polarity stability and adjust unstable features."""
-
-        adjustments: Dict[str, Dict[str, Any]] = {}
-        correlation_history: Dict[str, List[float]] = {}
-        sign_stability: Dict[str, float] = {}
-
-        if X.empty:
-            return adjustments, correlation_history, sign_stability
-
-        y_series = y if isinstance(y, pd.Series) else pd.Series(y, index=X.index)
-        try:
-            y_numeric = pd.to_numeric(y_series, errors='coerce')
-        except Exception:
-            y_numeric = pd.Series(y_series, index=y_series.index, dtype=float)
-
-        if not y_numeric.index.equals(X.index):
-            y_numeric = y_numeric.reindex(X.index)
-
-        valid_mask = ~y_numeric.isna()
-        X_numeric = X.select_dtypes(include=[np.number])
-
-        if X_numeric.empty:
-            return adjustments, correlation_history, sign_stability
-
-        X_numeric = X_numeric.loc[valid_mask]
-        y_valid = y_numeric.loc[valid_mask]
-
-        total_rows = len(X_numeric)
-        if total_rows < 15:
-            for col in X_numeric.columns:
-                sign_stability[col] = 1.0
-            return adjustments, correlation_history, sign_stability
-
-        window_size = max(total_rows // 5, 10)
-        window_size = min(window_size, total_rows - 1)
-        if window_size < 5:
-            for col in X_numeric.columns:
-                sign_stability[col] = 1.0
-            return adjustments, correlation_history, sign_stability
-
-        step_size = max(window_size // 2, 1)
-
-        correlation_accumulator: Dict[str, List[float]] = defaultdict(list)
-        sign_accumulator: Dict[str, List[int]] = defaultdict(list)
-
-        for start in range(0, total_rows - window_size + 1, step_size):
-            end = start + window_size
-            window_X = X_numeric.iloc[start:end]
-            window_y = y_valid.iloc[start:end]
-
-            if window_y.std(ddof=0) == 0:
-                continue
-
-            window_correlations = window_X.corrwith(window_y)
-
-            for feature, corr_value in window_correlations.items():
-                if pd.isna(corr_value):
-                    continue
-
-                corr_float = float(corr_value)
-                correlation_accumulator[feature].append(corr_float)
-
-                if abs(corr_float) < 1e-9:
-                    sign_accumulator[feature].append(0)
-                else:
-                    sign_accumulator[feature].append(1 if corr_float > 0 else -1)
-
-        recent_window_depth = 10
-        valid_index = y_valid.index
-
-        for feature in X_numeric.columns:
-            history = correlation_accumulator.get(feature, [])
-            correlation_history[feature] = history[-recent_window_depth:]
-
-            signs = sign_accumulator.get(feature, [])
-            recent_signs = signs[-recent_window_depth:]
-            non_zero_signs = [s for s in recent_signs if s != 0]
-
-            if len(non_zero_signs) <= 1:
-                flip_rate = 0.0
-            else:
-                flips = sum(
-                    non_zero_signs[idx] != non_zero_signs[idx - 1]
-                    for idx in range(1, len(non_zero_signs))
-                )
-                flip_rate = flips / (len(non_zero_signs) - 1)
-
-            stability_score = max(0.0, 1.0 - flip_rate)
-            sign_stability[feature] = stability_score
-
-            if flip_rate <= 0.7:
-                continue
-
-            dominant_sign = 0
-            if non_zero_signs:
-                positives = non_zero_signs.count(1)
-                negatives = non_zero_signs.count(-1)
-                if positives > negatives:
-                    dominant_sign = 1
-                elif negatives > positives:
-                    dominant_sign = -1
-
-            original_corr = X_numeric[feature].corr(y_valid)
-            final_corr = original_corr
-            action = 're_standardized'
-
-            if pd.notna(original_corr) and original_corr < 0:
-                X[feature] = -X[feature]
-                final_corr = X.loc[valid_index, feature].corr(y_valid)
-                action = 'inverted'
-            else:
-                std = X[feature].std(ddof=0)
-                if std and std > 0:
-                    mean = X[feature].mean()
-                    X[feature] = (X[feature] - mean) / std
-                    final_corr = X.loc[valid_index, feature].corr(y_valid)
-                else:
-                    action = 're_standardization_skipped_zero_variance'
-
-            adjustments[feature] = {
-                'action': action,
-                'flip_rate': float(flip_rate),
-                'sign_stability': float(stability_score),
-                'window_size': int(window_size),
-                'step_size': int(step_size),
-                'evaluated_windows': int(len(non_zero_signs)),
-                'dominant_sign': int(dominant_sign),
-                'original_correlation': None if pd.isna(original_corr) else float(original_corr),
-                'final_correlation': None if pd.isna(final_corr) else float(final_corr),
-                'correlation_history': [float(v) for v in correlation_history[feature]],
-                'recent_signs': non_zero_signs,
-                'reason': 'dominant_sign_flipped_gt_70_percent',
-                'timestamp': time.time(),
-            }
-
-        return adjustments, correlation_history, sign_stability
-
-    def _prepare_initial_features(self, X: pd.DataFrame, y: pd.Series, feature_names: Optional[List[str]] = None) -> pd.DataFrame:
-        """Prepare initial features for selection using vectorized operations."""
-
-        tprint("🔄 Preparing initial features with vectorized operations...")
-        tprint(f"📊 Input: {X.shape[0]} samples, {X.shape[1]} features")
-        
-        # Handle feature names
-        if feature_names is not None:
-            X = X[feature_names] if len(feature_names) <= len(X.columns) else X
-            tprint(f"📊 After feature name filtering: {X.shape[1]} features")
-        
-        # Remove low variance features using vectorized operations
-        tprint("🔄 Computing vectorized variance analysis...")
-        variance_array = self._vectorized_variance_analysis(X)
-        variance_threshold = self.config.min_variance_threshold
-        
-        low_variance_mask = variance_array < variance_threshold
-        low_variance_features = X.columns[low_variance_mask].tolist()
-        
-        if len(low_variance_features) > 0:
-            tprint(f"🗑️ Removing {len(low_variance_features)} low variance features (threshold: {variance_threshold})")
-            X = X.drop(columns=low_variance_features)
-            tprint(f"📊 After variance filtering: {X.shape[1]} features")
-        else:
-            tprint("✅ No low variance features to remove")
-
-        # Analyze rolling correlation polarity before correlation filtering
-        tprint("🔍 Evaluating rolling feature-target correlation stability...")
-        adjustments, correlation_history, sign_stability = self._evaluate_feature_polarity(X, y)
-        self.feature_polarity_adjustments = adjustments
-        self.feature_polarity_history = correlation_history
-        self.feature_sign_stability = sign_stability
-        self.results.polarity_adjustments = adjustments
-        self.results.sign_stability = sign_stability
-
-        if adjustments:
-            tprint(f"⚖️ Applied polarity adjustments to {len(adjustments)} features")
-        else:
-            tprint("✅ No polarity adjustments required")
-
-        # Apply gate feature protection before correlation filtering
-        if self.gate_protector:
-            tprint("🛡️ Applying gate-aware correlation filtering...")
-            X, protection_info = self.gate_protector.protect_gate_features(X, y, "correlation_filtering")
-            if protection_info.get('valid_gate_count', 0) > 0:
-                tprint(f"✅ Protected {protection_info['valid_gate_count']} gate features")
-            tprint(f"📊 After gate protection: {X.shape[1]} features")
-
-        # Remove highly correlated features using vectorized correlation computation
-        tprint("🔄 Computing vectorized correlation analysis...")
-        correlation_threshold = self.config.min_correlation_threshold
-        high_corr_features = self._find_highly_correlated_features_vectorized(X, correlation_threshold)
-
-        if len(high_corr_features) > 0:
-            tprint(f"🗑️ Removing {len(high_corr_features)} highly correlated features (threshold: {correlation_threshold})")
-            X = X.drop(columns=high_corr_features)
-            tprint(f"📊 After correlation filtering: {X.shape[1]} features")
-        else:
-            tprint("✅ No highly correlated features to remove")
-
-        # Apply entropy stability filtering before staging if enabled
-        if self.config.enable_entropy_balancing and len(X.columns) > 0:
-            tprint("🔄 Evaluating entropy stability across temporal slices...")
-            X = self._apply_entropy_balancing_filter(X)
-            tprint(f"📊 After entropy balancing: {X.shape[1]} features")
-        else:
-            tprint("⏭️ Entropy balancing disabled or no features available for evaluation")
-        
-        # Select top features if we have too many using vectorized operations
-        if len(X.columns) > self.config.initial_features:
-            tprint(f"📊 Selecting top {self.config.initial_features} features initially")
-            # Use vectorized variance-based selection for initial filtering
-            remaining_variance = self._vectorized_variance_analysis(X)
-            sorted_indices = np.argsort(remaining_variance)[::-1]
-            top_indices = sorted_indices[:self.config.initial_features]
-            top_features = X.columns[top_indices].tolist()
-            X = X[top_features]
-            tprint(f"📊 After top feature selection: {X.shape[1]} features")
-        
-        tprint(f"✅ Prepared {len(X.columns)} features for selection")
-        return X
-
-    def _apply_entropy_balancing_filter(self, X: pd.DataFrame) -> pd.DataFrame:
-        """Apply entropy stability filtering and record diagnostics."""
-
-        balancer_config = EntropyBalancerConfig(
-            num_slices=self.config.entropy_num_slices,
-            min_slice_size=self.config.entropy_min_slice_size,
-            max_entropy_variance=self.config.entropy_variance_threshold,
-            max_bins=self.config.entropy_max_bins,
-            min_unique_values=self.config.entropy_min_unique_values,
-            use_time_index=self.config.entropy_use_time_index,
-        )
-
-        filter_engine = EntropyStabilityFilter(balancer_config)
-        filter_result: EntropyFilterResult = filter_engine.filter(X)
-
-        self.results.entropy_variance = filter_result.entropy_variance
-        self.results.entropy_stability = filter_result.stability_scores
-        self.results.entropy_removed_features = filter_result.dropped_features
-
-        if filter_result.dropped_features:
-            dropped_preview = list(filter_result.dropped_features.items())[:5]
-            tprint(f"🗑️ Entropy filter removed {len(filter_result.dropped_features)} unstable features")
-            for feature, variance in dropped_preview:
-                tprint(f"   • {feature}: variance={variance:.4f}")
-            if filter_result.selected_features:
-                return X[filter_result.selected_features]
-
-            # Fallback: retain lowest variance features to avoid empty set
-            tprint("⚠️ All features flagged as unstable; retaining the most stable subset")
-            sorted_features = sorted(
-                filter_result.entropy_variance.items(),
-                key=lambda item: item[1]
-            )
-            fallback_count = max(1, int(len(sorted_features) * 0.5))
-            fallback_features = [feature for feature, _ in sorted_features[:fallback_count]]
-            self.results.entropy_removed_features = {
-                feature: variance
-                for feature, variance in filter_result.entropy_variance.items()
-                if feature not in fallback_features
-            }
-            return X[fallback_features]
-
-        tprint("✅ Entropy filter retained all features")
-        return X
-
-    def _find_highly_correlated_features_vectorized(self, X: pd.DataFrame, threshold: float) -> List[str]:
-        """Find and remove highly correlated features using vectorized correlation computation."""
-        tprint(f"🔄 Computing vectorized correlation matrix for {len(X.columns)} features...")
-        
-        # Use vectorized correlation computation
-        corr_matrix = self._vectorized_correlation_analysis(X)
-        
-        # Find features to drop using vectorized operations
-        corr_matrix_abs = np.abs(corr_matrix)
-        upper_triangle = np.triu(corr_matrix_abs, k=1)
-        
-        # Find features to drop
-        to_drop_mask = np.any(upper_triangle > threshold, axis=0)
-        to_drop = X.columns[to_drop_mask].tolist()
-        
-        tprint(f"📊 Correlation analysis: {len(to_drop)} features above threshold {threshold}")
-        return to_drop
-    
-    def _find_highly_correlated_features(self, X: pd.DataFrame, threshold: float) -> List[str]:
-        """Find and remove highly correlated features using optimized correlation computation."""
-        if MATRIX_OPERATIONS_AVAILABLE and len(X.columns) > 100:
-            # Use GPU-accelerated correlation for large datasets
-            self.logger.info(f"🚀 Using GPU-accelerated correlation for {len(X.columns)} features")
-            try:
-                corr_matrix = correlation_matrix_gpu(X)
-                corr_matrix = pd.DataFrame(corr_matrix, index=X.columns, columns=X.columns)
-            except Exception as e:
-                self.logger.warning(f"⚠️ GPU correlation failed, falling back to CPU: {e}")
-                corr_matrix = X.corr()
-        else:
-            corr_matrix = X.corr()
-
-        corr_matrix = corr_matrix.abs()
-        upper_triangle = corr_matrix.where(
-            np.triu(np.ones(corr_matrix.shape), k=1).astype(bool)
-        )
-
-        # Find features to drop
-        to_drop = [column for column in upper_triangle.columns if any(upper_triangle[column] > threshold)]
-
-        return to_drop
-
-
-
-    def _train_lightgbm_model(self, X: pd.DataFrame, y: pd.Series):
-        """Train LightGBM model with optimized parameters."""
-        tprint_debug("🚀 Training LightGBM model")
-        if not LIGHTGBM_AVAILABLE:
-            raise ImportError("LightGBM not available")
-
-        tprint(f"🚀 Training LightGBM model on {X.shape[0]} samples × {X.shape[1]} features")
-
-        # Determine if classification or regression
-        if self._is_classification(y):
-            objective = 'binary' if len(y.unique()) == 2 else 'multiclass'
-            params = self.config.lightgbm_params.copy()
-            params['objective'] = objective
-            params['num_class'] = len(y.unique()) if objective == 'multiclass' else 1
-
-            model = lgb.LGBMClassifier(**params)
-            tprint(f"🎯 LightGBM classification objective: {objective}")
-        else:
-            params = self.config.lightgbm_params.copy()
-            params['objective'] = 'regression'
-
-            model = lgb.LGBMRegressor(**params)
-            tprint("🎯 LightGBM regression objective")
-
-        # Train model
-        model.fit(X, y)
-        tprint("✅ LightGBM training complete")
-
-        return model
-
-    def _train_optimized_model(self, X: pd.DataFrame, y: pd.Series):
-        """Train either LightGBM or RandomForest based on availability and performance."""
-        tprint_debug("🚀 Training optimized model")
-        try:
-            if LIGHTGBM_AVAILABLE and len(X.columns) > 50:
-                self.logger.info("🚀 Using LightGBM for faster training")
-                tprint(f"🚀 Optimized training — selecting LightGBM (features: {len(X.columns)})")
-                return self._train_lightgbm_model(X, y)
-            else:
-                self.logger.info("📊 Using RandomForest (LightGBM not available or dataset too small)")
-                tprint("📊 Optimized training — selecting RandomForest")
-                return self._train_random_forest(X, y)
-        except Exception as e:
-            self.logger.info(f"📊 LightGBM training failed, falling back to RandomForest: {e}")
-            tprint(f"📊 LightGBM training failed ({e}), falling back to RandomForest")
-            return self._train_random_forest(X, y)
-
-
-
-    def _recursive_feature_elimination(self, X: pd.DataFrame, y: pd.Series) -> List[str]:
-        """Perform recursive feature elimination with early stopping."""
-        if not self.config.enable_rfe or not SKLEARN_FEATURE_SELECTION_AVAILABLE:
-            tprint("⏭️ RFE disabled or unavailable, keeping all features")
-            return X.columns.tolist()
-
-        # Apply gate feature protection before RFE
-        if self.gate_protector:
-            tprint("🛡️ Applying gate-aware RFE selection...")
-            X, protection_info = self.gate_protector.protect_gate_features(X, y, "rfe")
-            if protection_info.get('valid_gate_count', 0) > 0:
-                tprint(f"✅ Protected {protection_info['valid_gate_count']} gate features")
-            tprint(f"📊 After gate protection: {X.shape[1]} features")
-
-        current_features = X.columns.tolist()
-        best_score = float('-inf') if self._is_classification(y) else float('inf')
-        no_improvement_count = 0
-        step_count = 0
-
-        tprint(f"🔄 Starting RFE: {len(current_features)} features -> target: ~{self.config.rfe_min_features}")
-        tprint(f"📉 Step size: {self.config.rfe_step_size:.1%} ({max(1, int(len(current_features) * self.config.rfe_step_size))} features)")
-
-        while len(current_features) > self.config.rfe_min_features:
-            step_count += 1
-            # Calculate step size
-            step_size = max(1, int(len(current_features) * self.config.rfe_step_size))
-
-            if len(current_features) - step_size < self.config.rfe_min_features:
-                tprint(f"🛑 Stopping RFE: would go below minimum ({self.config.rfe_min_features})")
-                break
-
-            # Try removing step_size features
-            try:
-                tprint(f"🔄 RFE Step {step_count}: Training model on {len(current_features)} features...")
-                estimator = self._train_optimized_model(X[current_features], y)
-
-                tprint(f"🔄 RFE Step {step_count}: Selecting {len(current_features) - step_size} features...")
-                selector = RFE(estimator, n_features_to_select=len(current_features) - step_size, step=1)
-                selector = selector.fit(X[current_features], y)
-
-                selected_features = [f for f, selected in zip(current_features, selector.support_) if selected]
-
-                # Evaluate performance
-                if self._is_classification(y):
-                    score = selector.estimator_.score(X[selected_features], y)
-                else:
-                    score = -selector.estimator_.score(X[selected_features], y)  # Negative for regression
-
-                tprint(f"📊 RFE Step {step_count}: {len(selected_features)} features, score: {score:.4f}")
-
-                # Check for improvement
-                if (self._is_classification(y) and score > best_score) or \
-                   (not self._is_classification(y) and score < best_score):
-                    best_score = score
-                    current_features = selected_features
-                    no_improvement_count = 0
-                    tprint(f"✅ RFE Step {step_count}: Score improved to {score:.4f}")
-                else:
-                    no_improvement_count += 1
-                    tprint(f"⚠️ RFE Step {step_count}: No improvement (patience: {no_improvement_count}/{self.config.rfe_early_stopping_patience})")
-                    if no_improvement_count >= self.config.rfe_early_stopping_patience:
-                        tprint(f"🛑 RFE early stopping: no improvement for {no_improvement_count} steps")
-                        break
-
-            except Exception as e:
-                tprint(f"❌ RFE step {step_count} failed: {e}")
-                break
-
-        tprint(f"✅ RFE completed: {len(current_features)} features selected in {step_count} steps")
-        return current_features
-
-    def _process_features_in_chunks(self, X: pd.DataFrame, y: pd.Series) -> pd.DataFrame:
-        """Process features in chunks for memory efficiency."""
-        if not self.config.enable_chunked_processing or not MATRIX_OPERATIONS_AVAILABLE:
-            tprint("⏭️ Chunked processing disabled or unavailable")
-            return X
-
-        total_features = len(X.columns)
-        chunk_size = min(self.config.chunk_size, total_features)
-
-        if total_features <= chunk_size:
-            tprint(f"📊 Dataset small ({total_features} features), no chunking needed")
-            return X
-
-        tprint(f"📦 Processing {total_features} features in chunks of {chunk_size}")
-        tprint(f"🔗 Chunk overlap: {self.config.chunk_overlap} features")
-
-        # Split features into chunks with overlap
-        all_selected_features = []
-        chunk_overlap = self.config.chunk_overlap
-
-        for i in range(0, total_features, chunk_size - chunk_overlap):
-            start_idx = i
-            end_idx = min(i + chunk_size, total_features)
-
-            chunk_features = X.columns[start_idx:end_idx].tolist()
-            X_chunk = X[chunk_features]
-
-            tprint(f"🔄 Processing chunk {i//chunk_size + 1}: {len(chunk_features)} features")
-
-            # Process this chunk
-            chunk_selected = self._select_features_chunk(X_chunk, y)
-
-            # Add to overall selection
-            all_selected_features.extend(chunk_selected)
-
-            # Remove duplicates while preserving order
-            seen = set()
-            unique_features = []
-            for f in all_selected_features:
-                if f not in seen:
-                    seen.add(f)
-                    unique_features.append(f)
-
-            all_selected_features = unique_features
-
-            tprint(f"📈 Chunk result: {len(chunk_selected)} selected, total so far: {len(all_selected_features)}")
-
-            # Stop if we have enough features
-            if len(all_selected_features) >= self.config.target_features * 1.5:  # 50% buffer
-                tprint(f"🎯 Target reached ({len(all_selected_features)} features), stopping early")
-                break
-
-            # Limit total chunks
-            if (i // chunk_size) >= self.config.max_chunks - 1:
-                tprint(f"⚠️ Max chunks ({self.config.max_chunks}) reached, stopping")
-                break
-
-        tprint(f"📦 Chunked processing completed: {len(all_selected_features)} unique features selected")
-        return X[all_selected_features]
-
-    def _select_features_chunk(self, X_chunk: pd.DataFrame, y: pd.Series) -> List[str]:
-        """Select features for a single chunk."""
-        # Use a simplified version of the selection process for chunks
-        if len(X_chunk.columns) < 10:
-            tprint(f"📦 Chunk small ({len(X_chunk.columns)} features), selecting all")
-            return X_chunk.columns.tolist()
-
-        # Train model and get importance
-        try:
-            tprint(f"🚀 Training chunk model for {len(X_chunk.columns)} features")
-            model = self._train_optimized_model(X_chunk, y)
-            if hasattr(model, 'feature_importances_'):
-                importance = dict(zip(X_chunk.columns, model.feature_importances_))
-            else:
-                # For LightGBM, get feature importance
-                importance = dict(zip(X_chunk.columns, model.feature_importances_))
-
-            # Select top features
-            sorted_features = sorted(importance.items(), key=lambda x: x[1], reverse=True)
-            target_count = min(len(X_chunk.columns) // 2, self.config.target_features // 4)  # Conservative selection
-            selected = [f for f, _ in sorted_features[:target_count]]
-
-            tprint(f"✅ Chunk selection complete — {len(selected)} features retained")
-
-            return selected
-
-        except Exception as e:
-            self.logger.warning(f"⚠️ Chunk selection failed: {e}")
-            tprint(f"⚠️ Chunk selection failed ({e}), falling back to variance selection")
-            # Fallback to simple variance-based selection
-            variances = X_chunk.var().sort_values(ascending=False)
-            target_count = min(len(X_chunk.columns) // 2, 10)
-            fallback_features = variances.head(target_count).index.tolist()
-            tprint(f"✅ Variance fallback selected {len(fallback_features)} features")
-            return fallback_features
-    
-    def _stage_1_selection(self, X: pd.DataFrame, y: pd.Series, target_count: Optional[int] = None) -> Tuple[List[str], pd.Series]:
-        """Stage 1: Vectorized mRMR + Spearman with corrected math and argpartition selection."""
-        
-        # Sanitize input and initialize caches
-        X = self._sanitize_X(X)
-        
-        # Apply VectorBT memory optimization for stage 1
-        if VECTORBT_AVAILABLE and X.shape[1] > 100:
-            X = self._vectorbt_memory_optimization(X, "stage_1_selection")
-        
-        self._variance_cache = {}
-        self._X_cached = X
-        self._last_X = X
-        
-        p = X.shape[1]
-        t = target_count or 60
-        surplus = max(0, p - t)
-        n_keep = p - math.ceil(surplus / 2)
-        
-        tprint(f"🚀 Stage 1 (Vectorized): {p} → {n_keep} (remove {p - n_keep})")
-        
-        # Early termination check
-        if p < 10:
-            tprint_warning("⚠️ Stage 1: Insufficient features for selection")
-            return X.columns.tolist(), pd.Series(1.0, index=X.columns)
-        
-        try:
-            # Calculate scores vectorized
-            mrmr_scores = self._calculate_mrmr_mid_vectorized(X, y)
-            spearman_scores = self.spearman_abs_vectorized(X, y)
-            
-            # Early termination if scores are too low - use fallback selection
-            if self._early_termination_check(mrmr_scores.values, threshold=0.001):
-                tprint_warning("⚠️ Stage 1: Insufficient mRMR scores - using fallback selection")
-                fallback_features = X.columns.tolist()[:n_keep]
-                fallback_scores = pd.Series(0.5, index=fallback_features)
-                return fallback_features, fallback_scores
-            
-            if self._early_termination_check(spearman_scores.values, threshold=0.001):
-                tprint_warning("⚠️ Stage 1: Insufficient Spearman scores - using fallback selection")
-                fallback_features = X.columns.tolist()[:n_keep]
-                fallback_scores = pd.Series(0.5, index=fallback_features)
-                return fallback_features, fallback_scores
-            
-            # Vectorized z-score fusion with configurable weights
-            S = np.vstack([mrmr_scores.values, spearman_scores.values]).T  # (p, 2)
-            S_z = self.zscore_matrix(S, axis=0)
-            weights = self.config.stage_1_weights
-            combined_scores = weights['mrmr'] * S_z[:, 0] + weights['spearman'] * S_z[:, 1]  # (p,)
-            
-            # Early termination on combined scores - use fallback selection
-            if self._early_termination_check(combined_scores, threshold=0.001):
-                tprint_warning("⚠️ Stage 1: Insufficient combined scores - using fallback selection")
-                fallback_features = X.columns.tolist()[:n_keep]
-                fallback_scores = pd.Series(0.5, index=fallback_features)
-                return fallback_features, fallback_scores
-            
-            # Top-k selection with argpartition
-            selected = self.top_k(X.columns.tolist(), combined_scores, n_keep)
-            
-            # Validate selected features - use fallback if quality is insufficient
-            if not self._validate_feature_quality(X, y, selected):
-                tprint_warning("⚠️ Stage 1: Selected features failed quality validation - using fallback selection")
-                fallback_features = X.columns.tolist()[:n_keep]
-                fallback_scores = pd.Series(0.5, index=fallback_features)
-                return fallback_features, fallback_scores
-            
-            # Return scores as Series for consistency
-            scores = pd.Series(combined_scores, index=X.columns)
-            return selected, scores
-            
-        except Exception as e:
-            tprint_error(f"❌ Stage 1 selection failed: {e}")
-            # Fast fail with proper error instead of returning all features
-            raise RuntimeError(f"Stage 1 feature selection failed: {e}") from e
-
-    def _validate_inputs(self, X: pd.DataFrame, y: pd.Series) -> None:
-        """Comprehensive input validation with fast failing."""
-        # Fast fail on empty data
-        if X.empty:
-            raise ValueError("Feature matrix X cannot be empty")
-        if y.empty:
-            raise ValueError("Target vector y cannot be empty")
-        
-        # Fast fail on dimension mismatch
-        if len(X) != len(y):
-            raise ValueError(f"Dimension mismatch: X has {len(X)} samples, y has {len(y)} samples")
-        
-        # Fast fail on insufficient features
-        if X.shape[1] < 10:
-            raise ValueError(f"Insufficient features: {X.shape[1]} features, minimum 10 required")
-        
-        # Fast fail on insufficient samples
-        if X.shape[0] < 100:
-            raise ValueError(f"Insufficient samples: {X.shape[0]} samples, minimum 100 required")
-        
-        # Fast fail on all-NaN features
-        all_nan_features = X.isnull().all()
-        if all_nan_features.any():
-            nan_features = all_nan_features[all_nan_features].index.tolist()
-            raise ValueError(f"Features with all NaN values not allowed: {nan_features}")
-        
-        # Fast fail on non-numeric data
-        non_numeric_cols = X.select_dtypes(exclude=[np.number]).columns
-        if len(non_numeric_cols) > 0:
-            raise ValueError(f"Non-numeric features not allowed: {non_numeric_cols.tolist()}")
-        
-        # Fast fail on constant features
-        constant_features = X.nunique(dropna=False) <= 1
-        if constant_features.any():
-            const_features = constant_features[constant_features].index.tolist()
-            raise ValueError(f"Constant features not allowed: {const_features}")
-        
-        # Fast fail on infinite values
-        inf_features = np.isinf(X.values).any(axis=0)
-        if inf_features.any():
-            inf_cols = X.columns[inf_features].tolist()
-            raise ValueError(f"Features with infinite values not allowed: {inf_cols}")
-        
-        # Fast fail on target validation
-        if y.isnull().all():
-            raise ValueError("Target vector y cannot be all NaN")
-        
-        if np.isinf(y).any():
-            raise ValueError("Target vector y cannot contain infinite values")
-        
-        # Fast fail on target distribution issues
-        unique_values = y.nunique()
-        if unique_values <= 1:
-            raise ValueError(f"Target vector y has insufficient variation: {unique_values} unique values")
-        
-        # Check for extreme target imbalance (for classification tasks)
-        if unique_values <= 10:  # Likely classification
-            value_counts = y.value_counts()
-            min_count = value_counts.min()
-            max_count = value_counts.max()
-            if min_count < 5 or (max_count / min_count) > 100:
-                raise ValueError(f"Target vector y has extreme class imbalance: min={min_count}, max={max_count}")
-        
-        # Fast fail on index misalignment
-        if not X.index.equals(y.index):
-            raise ValueError("X and y must have identical indices")
-        
-        # Fast fail on temporal ordering issues (if datetime index)
-        if isinstance(X.index, pd.DatetimeIndex):
-            if not X.index.is_monotonic_increasing:
-                tprint_warning("⚠️ Datetime index is not monotonic - this may affect time series feature selection")
-        
-        tprint("✅ Input validation passed - all checks successful")
-
-    def _sanitize_X(self, X: pd.DataFrame) -> pd.DataFrame:
-        """Sanitize input data: numeric only, drop constants."""
-        # Validate inputs first
-        self._validate_inputs(X, pd.Series([0] * len(X), index=X.index))
-        
-        Xn = X.select_dtypes(include=[np.number]).copy()
-        nonconst = Xn.nunique(dropna=False) > 1
-        return Xn.loc[:, nonconst]
-    
-    def _stage_2_selection(self, X: pd.DataFrame, y: pd.Series, target_count: Optional[int] = None) -> Tuple[List[str], pd.Series]:
-        """Stage 2: Vectorized iterative bottom removal with boolean masks."""
-        try:
-            # Apply VectorBT memory optimization for stage 2
-            if VECTORBT_AVAILABLE and X.shape[1] > 50:
-                X = self._vectorbt_memory_optimization(X, "stage_2_selection")
-            
-            t = target_count or 60
-            target_plus_50 = t + 50
-            cols = X.columns.to_numpy()
-            mask = np.ones(len(cols), dtype=bool)  # True = keep
-            
-            # Initialize default scores
-            ensemble_scores = pd.Series(0.0, index=cols)
-            
-            tprint(f"🚀 Stage 2 (Vectorized): {len(cols)} → {target_plus_50} features")
-            
-            iteration = 0
-            max_iterations = 50  # Prevent infinite loops
-            
-            while np.sum(mask) > target_plus_50 and iteration < max_iterations:
-                iteration += 1
-                n_current = np.sum(mask)
-                current_cols = cols[mask]
-                
-                # Calculate ensemble scores with parallel CV
-                ensemble_scores = self.ensemble_scores_cv_parallel(X[current_cols], y, self.config.random_state)
-                
-                # Convert to array for argpartition
-                scores_array = ensemble_scores.values
-                
-                # Check for convergence (all scores too low)
-                if np.max(scores_array) < 0.001:
-                    tprint_warning("⚠️ Stage 2: All ensemble scores too low - stopping early")
-                    break
-                
-                # Remove bottom 20% (safely)
-                n_remove = max(1, min(n_current // 5, n_current - target_plus_50))
-                
-                # Get indices to remove using argpartition
-                bottom_indices = self.bottom_k(current_cols.tolist(), scores_array, n_remove)
-                drop_indices = np.where(np.isin(cols, bottom_indices))[0]
-                
-                # Update mask
-                mask[drop_indices] = False
-                
-                tprint(f"   Iteration {iteration}: {n_current} → {np.sum(mask)} (removed {n_remove})")
-            
-            if iteration >= max_iterations:
-                tprint_warning(f"⚠️ Stage 2: Maximum iterations ({max_iterations}) reached, using current selection")
-            
-            # Return final features and scores
-            final_features = cols[mask].tolist()
-            tprint(f"   ✅ Stage 2 completed: {len(final_features)} features selected")
-            return final_features, ensemble_scores
-            
-        except Exception as e:
-            tprint_error(f"❌ Stage 2 selection failed: {e}")
-            # Return a fallback selection instead of raising
-            tprint_warning("⚠️ Using fallback feature selection for stage 2")
-            fallback_features = X.columns.tolist()[:target_count or 60]
-            fallback_scores = pd.Series(0.5, index=fallback_features)
-            return fallback_features, fallback_scores
-
-    def _stage_3_selection(self, X: pd.DataFrame, y: pd.Series, target_count: Optional[int] = None) -> Tuple[List[str], pd.Series]:
-        """Stage 3: Vectorized chunked RFE with stability selection."""
-        try:
-            # Apply VectorBT memory optimization for stage 3
-            if VECTORBT_AVAILABLE and X.shape[1] > 30:
-                X = self._vectorbt_memory_optimization(X, "stage_3_selection")
-            
-            t = target_count or 60
-            target_plus_20 = t + 20
-            cols = X.columns.to_numpy()
-            mask = np.ones(len(cols), dtype=bool)  # True = keep
-            
-            # Initialize default scores
-            rfe_scores = pd.Series(0.0, index=cols)
-            
-            tprint(f"🚀 Stage 3 (Vectorized): {len(cols)} → {t} features")
-            
-            # Phase 1: Remove in chunks of 5 until target + 20
-            tprint("   Phase 1: Chunked removal (5 at a time)")
-            phase1_iterations = 0
-            max_phase1_iterations = 20
-            
-            while np.sum(mask) > target_plus_20 and phase1_iterations < max_phase1_iterations:
-                phase1_iterations += 1
-                n_current = np.sum(mask)
-                current_cols = cols[mask]
-                
-                # Calculate ensemble + stability scores
-                ens = self.ensemble_scores_cv_parallel(X[current_cols], y, self.config.random_state)
-                stab = self.stability_scores_vectorized(
-                    X[current_cols], y,
-                    scorer=lambda X_, y_: self.ensemble_scores_cv_parallel(X_, y_, self.config.random_state),
-                    rs=self.config.random_state
-                )
-                
-                # Check for convergence
-                if np.max(ens.values) < 0.001 or np.max(stab.values) < 0.001:
-                    tprint_warning("⚠️ Stage 3: Insufficient ensemble or stability scores - stopping early")
-                    break
-                
-                # Vectorized z-score blending with configurable weights
-                S = np.vstack([ens.values, stab.values]).T  # (p, 2)
-                S_z = self.zscore_matrix(S, axis=0)
-                weights = self.config.stage_3_weights
-                blended_scores = weights['ensemble'] * S_z[:, 0] + weights['stability'] * S_z[:, 1]  # (p,)
-                
-                # Remove based on blended scores
-                n_remove = min(5, n_current - target_plus_20)
-                bottom_indices = self.bottom_k(current_cols.tolist(), blended_scores, n_remove)
-                drop_indices = np.where(np.isin(cols, bottom_indices))[0]
-                
-                mask[drop_indices] = False
-                
-                tprint(f"     {n_current} → {np.sum(mask)} (removed {n_remove})")
-            
-            if phase1_iterations >= max_phase1_iterations:
-                raise ValueError(f"Stage 3 Phase 1: Maximum iterations ({max_phase1_iterations}) reached without convergence")
-            
-            # Phase 2: Remove one by one until final target
-            tprint("   Phase 2: One-by-one removal")
-            phase2_iterations = 0
-            max_phase2_iterations = 30
-            
-            while np.sum(mask) > t and phase2_iterations < max_phase2_iterations:
-                phase2_iterations += 1
-                n_current = np.sum(mask)
-                current_cols = cols[mask]
-                
-                # Use RFE with ensemble scoring
-                rfe_scores = self.ensemble_scores_cv_parallel(X[current_cols], y, self.config.random_state)
-                scores_array = rfe_scores.values
-                
-                # Check for convergence
-                if np.max(scores_array) < 0.001:
-                    tprint_warning("⚠️ Stage 3 Phase 2: All RFE scores too low - stopping early")
-                    break
-                
-                # Remove worst feature
-                worst_feature = self.bottom_k(current_cols.tolist(), scores_array, 1)[0]
-                drop_idx = np.where(cols == worst_feature)[0][0]
-                mask[drop_idx] = False
-                
-                tprint(f"     {n_current} → {np.sum(mask)} (removed {worst_feature})")
-            
-            if phase2_iterations >= max_phase2_iterations:
-                tprint_warning(f"⚠️ Stage 3 Phase 2: Maximum iterations ({max_phase2_iterations}) reached, using current selection")
-            
-            # Return final features and scores
-            final_features = cols[mask].tolist()
-            tprint(f"   ✅ Stage 3 completed: {len(final_features)} features selected")
-            return final_features, rfe_scores
-            
-        except Exception as e:
-            tprint_error(f"❌ Stage 3 selection failed: {e}")
-            # Return a fallback selection instead of raising
-            tprint_warning("⚠️ Using fallback feature selection for stage 3")
-            fallback_features = X.columns.tolist()[:target_count or 60]
-            fallback_scores = pd.Series(0.5, index=fallback_features)
-            return fallback_features, fallback_scores
-
-    def _prepare_cv_splitter(
-        self,
-        X: pd.DataFrame,
-        y: pd.Series,
-        event_timestamps: Optional[Union[pd.Series, pd.Index, np.ndarray, List[pd.Timestamp]]] = None,
-    ) -> Tuple[pd.DataFrame, pd.Series, Any]:
-        """Return time-ordered data and an appropriate cross-validation splitter."""
-
-        X_cv = X.copy()
-        y_cv = y.copy()
-        time_index: Optional[pd.DatetimeIndex] = None
-
-        if event_timestamps is not None:
-            if isinstance(event_timestamps, pd.Series):
-                aligned = event_timestamps
-                if not aligned.index.equals(X.index):
-                    aligned = aligned.reindex(X.index)
-                if aligned.isna().any():
-                    raise ValueError("Event timestamps missing for some samples during alignment.")
-                time_index = pd.DatetimeIndex(pd.to_datetime(aligned.values))
-            else:
-                event_index = pd.Index(event_timestamps)
-                if len(event_index) != len(X_cv):
-                    raise ValueError("Event timestamps length must match the number of samples.")
-                time_index = pd.DatetimeIndex(pd.to_datetime(event_index))
-        elif isinstance(X_cv.index, pd.DatetimeIndex):
-            time_index = pd.DatetimeIndex(X_cv.index)
-
-        if time_index is not None:
-            order = np.argsort(time_index.values)
-            if not np.array_equal(order, np.arange(len(time_index))):
-                X_cv = X_cv.iloc[order].copy()
-                y_cv = y_cv.iloc[order].copy()
-                time_index = time_index[order]
-            else:
-                X_cv = X_cv.copy()
-                y_cv = y_cv.copy()
-
-            X_cv.index = time_index
-            if isinstance(y_cv, pd.Series):
-                y_cv.index = time_index
-        else:
-            X_cv = X_cv.copy()
-            y_cv = y_cv.copy()
-
-        label_horizon = int(getattr(self.config, 'label_horizon_minutes', 0) or 0)
-        purge_config = self.config.purge_minutes if self.config.purge_minutes is not None else label_horizon
-        embargo_config = self.config.embargo_minutes if self.config.embargo_minutes is not None else label_horizon
-        purge_minutes = max(label_horizon, int(purge_config))
-        embargo_minutes = max(label_horizon, int(embargo_config))
-
-        if time_index is not None and PURGED_KFOLD_AVAILABLE:
-            purge_delta = pd.Timedelta(minutes=purge_minutes)
-            embargo_delta = pd.Timedelta(minutes=embargo_minutes)
-            cv = PurgedKFoldTime(
-                n_splits=self.config.cv_folds,
-                purge=purge_delta,
-                embargo=embargo_delta,
-            )
-        else:
-            if self._is_classification(y_cv):
-                cv = StratifiedKFold(n_splits=self.config.cv_folds, shuffle=False)
-            else:
-                cv = KFold(n_splits=self.config.cv_folds, shuffle=False)
-
-        return X_cv, y_cv, cv
-
-    def _cross_validate_feature_importance_optimized(
-        self,
-        X: pd.DataFrame,
-        y: pd.Series,
-        event_timestamps: Optional[Union[pd.Series, pd.Index, np.ndarray, List[pd.Timestamp]]] = None,
-    ) -> Dict[str, float]:
-        """Cross-validation feature importance using optimized models."""
-        cv_scores = {}
-
-        try:
-            X_cv, y_cv, cv = self._prepare_cv_splitter(X, y, event_timestamps)
-
-            for fold, (train_idx, val_idx) in enumerate(cv.split(X_cv, y_cv)):
-                X_train, X_val = X_cv.iloc[train_idx], X_cv.iloc[val_idx]
-                y_train, y_val = y_cv.iloc[train_idx], y_cv.iloc[val_idx]
-
-                # Train optimized model on fold
-                model = self._train_optimized_model(X_train, y_train)
-
-                # Get feature importance from this fold
-                if hasattr(model, 'feature_importances_'):
-                    fold_importance = dict(zip(X_cv.columns, model.feature_importances_))
-                else:
-                    # Fallback to variance-based importance
-                    fold_importance = dict(zip(X_cv.columns, X_train.var().values))
-
-                # Accumulate scores
-                for feature, importance in fold_importance.items():
-                    if feature not in cv_scores:
-                        cv_scores[feature] = []
-                    cv_scores[feature].append(importance)
-
-            # Average across folds
-            for feature in cv_scores:
-                cv_scores[feature] = np.mean(cv_scores[feature])
-
-        except Exception as e:
-            self.logger.warning(f"⚠️ Optimized CV feature importance failed: {e}")
-            # Fallback to simple variance-based scores
-            cv_scores = dict(zip(X.columns, X.var().values))
-
-        return cv_scores
-    
-    def _shap_based_selection(self, X: pd.DataFrame, y: pd.Series, target_count: int) -> Tuple[List[str], Dict[str, float]]:
-        """SHAP-based feature selection."""
-
-        # Sample data for SHAP analysis
-        sample_size = min(self.config.shap_sample_size, len(X))
-        sample_indices = np.random.choice(len(X), sample_size, replace=False)
-        X_sample = X.iloc[sample_indices]
-        y_sample = y.iloc[sample_indices]
-        tprint(f"🔮 SHAP analysis on sample of {sample_size} observations")
-
-        # Train model
-        rf_model = self._train_random_forest(X_sample, y_sample)
-        tprint("🌳 RandomForest trained for SHAP analysis")
-
-        # Calculate SHAP values
-        explainer = shap.TreeExplainer(rf_model)
-        shap_values = explainer.shap_values(X_sample)
-        tprint("✨ SHAP values computed")
-
-        # Calculate mean absolute SHAP values
-        if isinstance(shap_values, list):  # Classification
-            shap_importance = np.mean(np.abs(shap_values), axis=(0, 1))
-        else:  # Regression
-            shap_importance = np.mean(np.abs(shap_values), axis=0)
-
-        # Create feature importance dictionary
-        shap_importance_dict = dict(zip(X.columns, shap_importance))
-
-        # Select top features
-        sorted_features = sorted(shap_importance_dict.items(), key=lambda x: x[1], reverse=True)
-        selected_features = [f[0] for f in sorted_features[:target_count]]
-        tprint(f"✅ SHAP selected top {len(selected_features)} features")
-
-        # Calculate scores
-        scores = {
-            'shap_importance_score': np.mean(shap_importance),
-            'shap_variance': np.var(shap_importance),
-            'selection_confidence': len(selected_features) / len(X.columns)
-        }
-
-        tprint(
-            f"📊 SHAP scores — mean importance: {scores['shap_importance_score']:.6f},"
-            f" variance: {scores['shap_variance']:.6f}"
-        )
-
-        return selected_features, scores
-
-    def _enhanced_rf_selection(self, X: pd.DataFrame, y: pd.Series, target_count: int) -> Tuple[List[str], Dict[str, float]]:
-        """Enhanced RandomForest selection with multiple criteria."""
-
-        # Train multiple RandomForest models with different parameters
-        models = []
-        tprint("🌳 Starting enhanced RandomForest ensemble selection")
-        for n_est in [50, 100, 150]:
-            model = RandomForestRegressor(
-                n_estimators=n_est,
-                max_depth=self.config.rf_max_depth,
-                min_samples_split=self.config.rf_min_samples_split,
-                random_state=self.config.rf_random_state + n_est
-            )
-            model.fit(X, y)
-            models.append(model)
-            tprint(f"✅ Trained RandomForest with {n_est} trees")
-
-        # Average feature importance across models
-        avg_importance = np.zeros(len(X.columns))
-        for model in models:
-            avg_importance += model.feature_importances_
-        avg_importance /= len(models)
-        tprint("📊 Averaged feature importance across ensemble")
-
-        # Create feature importance dictionary
-        importance_dict = dict(zip(X.columns, avg_importance))
-
-        # Select top features
-        sorted_features = sorted(importance_dict.items(), key=lambda x: x[1], reverse=True)
-        selected_features = [f[0] for f in sorted_features[:target_count]]
-        tprint(f"🏆 Enhanced RF selected {len(selected_features)} features")
-
-        # Calculate scores
-        scores = {
-            'enhanced_rf_score': np.mean(avg_importance),
-            'model_agreement': 1 - np.std(avg_importance) / np.mean(avg_importance),
-            'selection_quality': len(selected_features) / len(X.columns)
-        }
-
-        tprint(
-            f"📈 Enhanced RF scores — mean: {scores['enhanced_rf_score']:.6f},"
-            f" agreement: {scores['model_agreement']:.6f}"
-        )
-
-        return selected_features, scores
-    
-    def _cross_validate_feature_importance(
-        self,
-        X: pd.DataFrame,
-        y: pd.Series,
-        event_timestamps: Optional[Union[pd.Series, pd.Index, np.ndarray, List[pd.Timestamp]]] = None,
-    ) -> Dict[str, float]:
-        """Cross-validate feature importance using multiple folds."""
-
-        cv_scores = {feature: 0.0 for feature in X.columns}
-
-        X_cv, y_cv, cv = self._prepare_cv_splitter(X, y, event_timestamps)
-
-        for fold, (train_idx, val_idx) in enumerate(cv.split(X_cv, y_cv)):
-            X_train, X_val = X_cv.iloc[train_idx], X_cv.iloc[val_idx]
-            y_train, y_val = y_cv.iloc[train_idx], y_cv.iloc[val_idx]
-
-            # Train model on fold
-            model = self._train_random_forest(X_train, y_train)
-
-            # Get feature importance
-            fold_importance = dict(zip(X_cv.columns, model.feature_importances_))
-
-            # Accumulate scores
-            for feature, importance in fold_importance.items():
-                cv_scores[feature] += importance
-
-        # Average across folds
-        for feature in cv_scores:
-            cv_scores[feature] /= self.config.cv_folds
-
-        return cv_scores
-    
-    def _train_random_forest(self, X: pd.DataFrame, y: pd.Series):
-        """Train RandomForest model."""
-        tprint_debug("🌲 Training RandomForest model")
-        if self._is_classification(y):
-            model = RandomForestClassifier(
-                n_estimators=self.config.rf_n_estimators,
-                max_depth=self.config.rf_max_depth,
-                min_samples_split=self.config.rf_min_samples_split,
-                random_state=self.config.rf_random_state
-            )
-            tprint(f"🌳 Training RandomForestClassifier with {self.config.rf_n_estimators} trees")
-        else:
-            model = RandomForestRegressor(
-                n_estimators=self.config.rf_n_estimators,
-                max_depth=self.config.rf_max_depth,
-                min_samples_split=self.config.rf_min_samples_split,
-                random_state=self.config.rf_random_state
-            )
-            tprint(f"🌳 Training RandomForestRegressor with {self.config.rf_n_estimators} trees")
-
-        model.fit(X, y)
-        tprint("✅ RandomForest training complete")
-        return model
-
-    def _predict_signal(self, model: Any, X: pd.DataFrame) -> pd.Series:
-        """Generate a continuous trading signal from a fitted model."""
-
-        if isinstance(X, pd.DataFrame):
-            index = X.index
-            data = X
-        else:
-            index = pd.RangeIndex(len(X))
-            data = X
-
-        # Prefer probabilistic outputs for classification problems to obtain a signal
-        try:
-            if hasattr(model, "predict_proba"):
-                proba = model.predict_proba(data)
-                if isinstance(proba, list):  # Some estimators may return list of arrays
-                    proba = np.asarray(proba)
-                if hasattr(proba, "ndim") and proba.ndim == 2 and proba.size:
-                    if proba.shape[1] == 2:
-                        signal = proba[:, 1] - proba[:, 0]
-                    else:
-                        # Use expectation over classes as a continuous proxy
-                        classes = getattr(model, "classes_", np.arange(proba.shape[1]))
-                        signal = proba @ np.asarray(classes, dtype=float)
-                    return pd.Series(signal, index=index, dtype=float)
-        except Exception as e:
-            tprint_warning(f"⚠️ Failed to get probability predictions: {e}")
-            pass
-
-        try:
-            predictions = model.predict(data)
-        except Exception:
-            predictions = model.predict(np.asarray(data))
-
-        return pd.Series(np.asarray(predictions, dtype=float), index=index)
-
-    def _evaluate_trading_metrics(self, y_true: pd.Series, predictions: pd.Series) -> Dict[str, float]:
-        """Compute trading-aware validation metrics from predictions."""
-
-        if y_true is None or predictions is None:
-            return {}
-
-        y_series = y_true if isinstance(y_true, pd.Series) else pd.Series(y_true)
-        pred_series = predictions if isinstance(predictions, pd.Series) else pd.Series(predictions)
-
-        data = pd.concat([y_series.rename("target"), pred_series.rename("prediction")], axis=1)
-        data = data.dropna()
-
-        if data.empty:
-            return {
-                'information_coefficient': 0.0,
-                'long_short_sharpe': 0.0,
-                'turnover': 0.0,
-                'mean_return': 0.0,
-                'std_return': 0.0,
-                'net_mean_return': 0.0,
-                'mean_transaction_cost': 0.0,
-                'transaction_cost': self.config.trading_cost,
-                'turnover_penalty_applied': 0.0,
-                'annualization_factor': float(np.sqrt(max(self.config.trading_horizon, 1))),
-                'n_observations': 0,
-            }
-
-        y_values = data['target'].to_numpy(dtype=float)
-        pred_values = data['prediction'].to_numpy(dtype=float)
-
-        ic_value = 0.0
-        if len(data) >= 2:
-            method = (self.config.ic_method or 'spearman').lower()
-            if method == 'spearman':
-                try:
-                    from scipy.stats import spearmanr  # type: ignore
-                    ic_value = float(spearmanr(pred_values, y_values)[0])
-                except Exception:
-                    ic_value = float(pd.Series(pred_values).rank().corr(pd.Series(y_values).rank()))
-            else:
-                ic_value = float(np.corrcoef(pred_values, y_values)[0, 1])
-            if not np.isfinite(ic_value):
-                ic_value = 0.0
-
-        positions = pd.Series(np.sign(pred_values), index=data.index, dtype=float)
-        position_changes = positions.diff().abs()
-        if not position_changes.empty:
-            position_changes.iloc[0] = abs(positions.iloc[0])
-
-        gross_returns = positions * data['target']
-        costs = position_changes * self.config.trading_cost
-        net_returns_before_impact = gross_returns - costs
-
-        turnover_metrics = calculate_turnover_metrics(
-            positions,
-            gross_returns,
-            position_changes=position_changes,
-            periods_per_year=float(self.config.trading_horizon or 252),
-        )
-
-        mean_transaction_cost = float(costs.mean()) if not costs.empty else 0.0
-        mean_return = float(gross_returns.mean()) if not gross_returns.empty else 0.0
-        net_mean_return = float(net_returns_before_impact.mean()) if not net_returns_before_impact.empty else 0.0
-
-        volume_series = getattr(self, 'market_volume_series', None)
-        if isinstance(volume_series, pd.Series):
-            volume_series = volume_series.reindex(data.index)
-        elif volume_series is not None:
-            volume_series = pd.Series(volume_series, index=data.index, dtype=float)
-
-        impact_adjusted_returns, market_impact_costs, capacity_exceeded, max_trade_size = apply_market_impact_model(
-            net_returns_before_impact,
-            position_changes,
-            volume=volume_series,
-            impact_coefficient=getattr(self.config, 'market_impact_coefficient', 0.1),
-            capacity_limit_usd=getattr(self.config, 'capacity_limit_usd', 1_000_000.0),
-        )
-
-        impact_adjusted_mean = float(impact_adjusted_returns.mean()) if not impact_adjusted_returns.empty else 0.0
-        std_return = float(impact_adjusted_returns.std(ddof=0)) if not impact_adjusted_returns.empty else 0.0
-        mean_market_impact = float(market_impact_costs.mean()) if not market_impact_costs.empty else 0.0
-        total_market_impact = float(market_impact_costs.sum()) if not market_impact_costs.empty else 0.0
-
-        penalty_adjustment = self.config.turnover_penalty * turnover_metrics['turnover_per_period']
-        adjusted_net_mean = impact_adjusted_mean - penalty_adjustment
-
-        horizon = max(self.config.trading_horizon, 1)
-        annualization_factor = float(np.sqrt(horizon))
-        sharpe = 0.0
-        if std_return > 1e-12:
-            sharpe = adjusted_net_mean / std_return * annualization_factor
-
-        rejected, rejection_reason = reject_high_turnover_configs(
-            {
-                'turnover_annual': turnover_metrics['turnover_annual'],
-                'sharpe': sharpe,
-            },
-            max_turnover_annual=getattr(self.config, 'max_turnover_annual', 50.0),
-            max_sharpe_to_turnover_ratio=getattr(self.config, 'min_sharpe_to_turnover_ratio', 0.1),
-        )
-
-        if rejected and rejection_reason:
-            tprint_warning(f"🚫 {rejection_reason}")
-
-        metrics: Dict[str, Any] = {
-            'information_coefficient': ic_value,
-            'long_short_sharpe': float(sharpe),
-            'turnover': turnover_metrics['turnover_per_period'],
-            'turnover_per_period': turnover_metrics['turnover_per_period'],
-            'turnover_annual': turnover_metrics['turnover_annual'],
-            'avg_holding_period_bars': turnover_metrics['avg_holding_period_bars'],
-            'position_stability': turnover_metrics['position_stability'],
-            'mean_return': mean_return,
-            'std_return': std_return,
-            'net_mean_return': net_mean_return,
-            'impact_adjusted_mean_return': impact_adjusted_mean,
-            'adjusted_mean_return': adjusted_net_mean,
-            'mean_transaction_cost': mean_transaction_cost,
-            'total_transaction_cost': float(costs.sum()) if not costs.empty else 0.0,
-            'mean_market_impact_cost': mean_market_impact,
-            'total_market_impact_cost': total_market_impact,
-            'transaction_cost': self.config.trading_cost,
-            'turnover_penalty_applied': penalty_adjustment,
-            'annualization_factor': annualization_factor,
-            'n_observations': int(len(data)),
-            'capacity_exceeded': capacity_exceeded,
-            'max_trade_size': max_trade_size,
-            'is_turnover_rejected': rejected,
-        }
-
-        if rejection_reason:
-            metrics['turnover_rejection_reason'] = rejection_reason
-
-        return metrics
-
-    def _generate_cv_trading_metrics(self, X: pd.DataFrame, y: pd.Series) -> Tuple[Dict[str, float], pd.Series, List[Dict[str, float]]]:
-        """Run cross-validation to obtain validation predictions and trading metrics."""
-
-        if y is None or len(y) == 0 or X is None or len(X) == 0:
-            empty_predictions = pd.Series(dtype=float)
-            return ({
-                'information_coefficient': 0.0,
-                'long_short_sharpe': 0.0,
-                'turnover': 0.0,
-                'mean_return': 0.0,
-                'std_return': 0.0,
-                'net_mean_return': 0.0,
-                'adjusted_mean_return': 0.0,
-                'mean_transaction_cost': 0.0,
-                'transaction_cost': self.config.trading_cost,
-                'turnover_penalty_applied': 0.0,
-                'annualization_factor': float(np.sqrt(max(self.config.trading_horizon, 1))),
-                'n_observations': 0,
-                'cv_folds': 0,
-                'fold_metrics': [],
-                'cv_mean': 0.0,
-                'cv_std': 0.0,
-            }, empty_predictions, [])
-
-        if not isinstance(y, pd.Series):
-            if isinstance(X, pd.DataFrame):
-                y_series = pd.Series(y, index=X.index)
-            else:
-                y_series = pd.Series(y)
-        else:
-            y_series = y
-
-        n_samples = len(y_series)
-        n_splits = min(self.config.cv_folds, n_samples)
-        if n_splits < 2:
-            model = self._train_random_forest(X, y_series)
-            predictions = self._predict_signal(model, X)
-            metrics = self._evaluate_trading_metrics(y_series, predictions)
-            metrics.update({
-                'cv_folds': 1,
-                'fold_metrics': [],
-                'cv_mean': metrics.get('long_short_sharpe', 0.0),
-                'cv_std': 0.0,
-            })
-            return metrics, predictions, []
-
-        if self._is_classification(y_series):
-            try:
-                cv_splitter = StratifiedKFold(n_splits=n_splits, shuffle=False)
-                split_iter = cv_splitter.split(X, y_series)
-            except ValueError:
-                cv_splitter = KFold(n_splits=n_splits, shuffle=False)
-                split_iter = cv_splitter.split(X, y_series)
-        else:
-            cv_splitter = KFold(n_splits=n_splits, shuffle=False)
-            split_iter = cv_splitter.split(X, y_series)
-
-        predictions = pd.Series(index=y_series.index, dtype=float)
-        fold_metrics: List[Dict[str, float]] = []
-
-        for fold_idx, (train_idx, val_idx) in enumerate(split_iter, start=1):
-            X_train = X.iloc[train_idx]
-            y_train = y_series.iloc[train_idx]
-            X_val = X.iloc[val_idx]
-            y_val = y_series.iloc[val_idx]
-
-            model = self._train_random_forest(X_train, y_train)
-            fold_predictions = self._predict_signal(model, X_val)
-            predictions.iloc[val_idx] = fold_predictions.values
-
-            fold_metric = self._evaluate_trading_metrics(y_val, fold_predictions)
-            fold_metric['fold'] = fold_idx
-            fold_metrics.append(fold_metric)
-
-        valid_mask = predictions.notna()
-        aggregated_metrics = self._evaluate_trading_metrics(y_series.loc[valid_mask], predictions.loc[valid_mask])
-
-        sharpe_values = [float(m.get('long_short_sharpe', 0.0)) for m in fold_metrics]
-        ic_values = [float(m.get('information_coefficient', 0.0)) for m in fold_metrics]
-        turnover_values = [float(m.get('turnover', 0.0)) for m in fold_metrics]
-
-        trading_cv_mean = float(np.mean(sharpe_values)) if sharpe_values else aggregated_metrics.get('long_short_sharpe', 0.0)
-        trading_cv_std = float(np.std(sharpe_values)) if sharpe_values else 0.0
-
-        aggregated_metrics.update({
-            'cv_folds': len(fold_metrics),
-            'fold_metrics': fold_metrics,
-            'cv_mean_trading': trading_cv_mean,
-            'cv_std_trading': trading_cv_std,
-            'cv_mean': trading_cv_mean,
-            'cv_std': trading_cv_std,
-            'sharpe_by_fold': sharpe_values,
-            'information_coefficient_by_fold': ic_values,
-            'turnover_by_fold': turnover_values,
-        })
-
-        return aggregated_metrics, predictions, fold_metrics
-
-    def _is_classification(self, y: pd.Series) -> bool:
-        """Determine if target is classification or regression."""
-        # Simple heuristic: if target has few unique values, treat as classification
-        unique_values = len(y.unique())
-        return unique_values <= 10 or y.dtype == 'category'
-    
-    def _calculate_agreement(self, scores1: Dict[str, float], scores2: Dict[str, float]) -> float:
-        """Calculate agreement between two scoring methods."""
-        common_features = set(scores1.keys()) & set(scores2.keys())
-        if not common_features:
-            return 0.0
-
-        # Calculate correlation between scores
-        scores1_values = [scores1[f] for f in common_features]
-        scores2_values = [scores2[f] for f in common_features]
-
-        correlation = np.corrcoef(scores1_values, scores2_values)[0, 1]
-        return correlation if not np.isnan(correlation) else 0.0
-
-    def _collect_validation_predictions(
-        self,
-        model: Any,
-        X: pd.DataFrame,
-        y: pd.Series,
-        features: List[str],
-        is_classification: bool,
-    ) -> Tuple[Optional[np.ndarray], Optional[np.ndarray], Optional[np.ndarray]]:
-        """Collect cross-validated predictions and probabilities for calibration analysis."""
-
-        if not self.config.enable_uncertainty_reporting:
-            return None, None, None
-
-        try:
-            selected_features = X[features] if features else X
-            X_values = selected_features.values
-            y_array = y.values if hasattr(y, 'values') else np.asarray(y)
-
-            if is_classification:
-                cv = StratifiedKFold(n_splits=self.config.cv_folds, shuffle=False)
-                classes = getattr(model, 'classes_', None)
-                if classes is None:
-                    classes = np.unique(y_array)
-                classes = np.asarray(classes)
-                predictions: np.ndarray = np.empty(len(y_array), dtype=object)
-                probabilities: Optional[np.ndarray] = np.zeros((len(y_array), len(classes)), dtype=float)
-            else:
-                cv = KFold(n_splits=self.config.cv_folds, shuffle=False)
-                predictions = np.zeros(len(y_array), dtype=float)
-                probabilities = None
-                classes = None
-
-            for train_idx, val_idx in cv.split(X_values, y_array):
-                model_clone = clone(model)
-                model_clone.fit(X_values[train_idx], y_array[train_idx])
-                fold_predictions = model_clone.predict(X_values[val_idx])
-                predictions[val_idx] = fold_predictions
-
-                if is_classification and probabilities is not None:
-                    if hasattr(model_clone, 'predict_proba'):
-                        fold_probabilities = model_clone.predict_proba(X_values[val_idx])
-                        if fold_probabilities.ndim == 1:
-                            fold_probabilities = np.column_stack([1 - fold_probabilities, fold_probabilities])
-                        fold_classes = getattr(model_clone, 'classes_', None)
-                        if fold_classes is None:
-                            fold_classes = classes
-                        fold_classes = np.asarray(fold_classes)
-                        aligned = np.zeros((len(val_idx), len(classes)), dtype=float)
-                        class_index_map = {cls: idx for idx, cls in enumerate(fold_classes)}
-                        for target_idx, cls in enumerate(classes):
-                            if cls in class_index_map:
-                                aligned[:, target_idx] = fold_probabilities[:, class_index_map[cls]]
-                        probabilities[val_idx] = aligned
-                    else:
-                        fallback = np.zeros((len(val_idx), len(classes)), dtype=float)
-                        class_index_map = {cls: idx for idx, cls in enumerate(classes)}
-                        for row_idx, label in enumerate(fold_predictions):
-                            cls_idx = class_index_map.get(label)
-                            if cls_idx is not None:
-                                fallback[row_idx, cls_idx] = 1.0
-                        probabilities[val_idx] = fallback
-
-            return predictions, probabilities, classes
-        except Exception as exc:
-            self.logger.warning(f"⚠️ Failed to collect CV predictions for calibration: {exc}")
-            return None, None, None
-    
-    def _compile_results(self,
-                        X: pd.DataFrame,
-                        y: pd.Series,
-                        stage_1_features: List[str],
-                        stage_2_features: List[str],
-                        stage_3_features: List[str],
-                        stage_1_scores: Union[Dict[str, Any], pd.Series],
-                        stage_2_scores: Union[Dict[str, Any], pd.Series],
-                        stage_3_scores: Union[Dict[str, Any], pd.Series]):
-        """Compile final results with support for Series scores."""
-        
-        is_classification = self._is_classification(y)
-
-        # Store feature lists
-        self.results.stage_1_features = stage_1_features
-        self.results.stage_2_features = stage_2_features
-        self.results.stage_3_features = stage_3_features
-        self.results.final_features = stage_3_features
-
-        # Convert Series scores to dict if needed
-        def _ensure_dict_scores(scores):
-            if isinstance(scores, pd.Series):
-                return scores.to_dict()
-            return scores
-
-        # Store scores
-        self.results.stage_1_scores = _ensure_dict_scores(stage_1_scores)
-        self.results.stage_2_scores = _ensure_dict_scores(stage_2_scores)
-        self.results.stage_3_scores = _ensure_dict_scores(stage_3_scores)
-
-        # Calculate final model performance
-        final_model = self._train_random_forest(X[stage_3_features], y)
-        trading_metrics, validation_predictions, fold_metrics = self._generate_cv_trading_metrics(X[stage_3_features], y)
-        model_score = float(final_model.score(X[stage_3_features], y)) if hasattr(final_model, 'score') else 0.0
-
-        trading_metrics = dict(trading_metrics)
-        trading_metrics['model_score'] = model_score
-        trading_metrics.setdefault('validation_sample_count', trading_metrics.get('n_observations', len(y)))
-        trading_metrics.setdefault('cv_mean_trading', trading_metrics.get('cv_mean'))
-        trading_metrics.setdefault('cv_std_trading', trading_metrics.get('cv_std'))
-        
-        try:
-            from src.utils.ml_common.validation.unified_cv import perform_cross_validation as unified_perform_cv
-            cv_res = unified_perform_cv(final_model, X[stage_3_features].values, y.values if hasattr(y, 'values') else y, cv_folds=self.config.cv_folds, scoring=self.config.cv_scoring)
-            cv_scores = np.array(cv_res.get('scores', []) or [])
-        except Exception:
-            cv_scores = np.array([])
-
-        cv_mean = float(np.nanmean(cv_scores)) if cv_scores.size else float('nan')
-        cv_std = float(np.nanstd(cv_scores)) if cv_scores.size else float('nan')
-        metric_name, metric_value = self._calculate_model_metric(final_model, X[stage_3_features], y)
-
-        final_scores: Dict[str, Any] = dict(trading_metrics)
-        final_scores.update({
-            'cv_metric': self.config.cv_scoring,
-            'cv_mean': cv_mean,
-            'cv_std': cv_std,
-            'model_cv_mean': cv_mean,
-            'model_cv_std': cv_std,
-            'cv_mean_trading': trading_metrics.get('cv_mean_trading'),
-            'cv_std_trading': trading_metrics.get('cv_std_trading'),
-            'validation_sample_count': trading_metrics.get('validation_sample_count'),
-            'model_score': model_score,
-        })
-
-        if metric_value is not None:
-            final_scores[metric_name] = metric_value
-
-        if final_scores.get('is_turnover_rejected'):
-            final_scores['eligible_for_selection'] = False
-            final_scores['cv_mean'] = float('nan')
-            final_scores['cv_std'] = float('nan')
-            final_scores['model_cv_mean'] = float('nan')
-            final_scores['model_cv_std'] = float('nan')
-        else:
-            final_scores['eligible_for_selection'] = True
-
-        self.results.final_scores = final_scores
-        self.results.eligible_for_selection = bool(final_scores.get('eligible_for_selection', True))
-        self.results.turnover_rejection_reason = final_scores.get('turnover_rejection_reason')
-
-        validation_predictions: Optional[np.ndarray]
-        validation_probabilities: Optional[np.ndarray]
-        validation_classes: Optional[np.ndarray]
-        validation_predictions = None
-        validation_probabilities = None
-        validation_classes = None
-        calibration_metrics: Dict[str, Any] = {}
-        reliability_diagram: Dict[str, Any] = {}
-        uncertainty_metrics: Dict[str, Any] = {}
-        classes_list: Optional[List[Any]] = None
-
-        if self.config.enable_uncertainty_reporting:
-            validation_predictions, validation_probabilities, validation_classes = self._collect_validation_predictions(
-                final_model,
-                X,
-                y,
-                stage_3_features,
-                is_classification,
-            )
-
-            if is_classification and validation_probabilities is not None and validation_classes is not None:
-                try:
-                    y_array = y.values if hasattr(y, 'values') else np.asarray(y)
-                    classes_list = validation_classes.tolist() if hasattr(validation_classes, 'tolist') else list(validation_classes)
-                    calibration_report = compute_classification_calibration(
-                        y_array,
-                        validation_probabilities,
-                        classes=classes_list,
-                        n_bins=max(2, int(self.config.reliability_bins)),
-                    )
-                    reliability_diagram = calibration_report.pop('reliability_diagram', {})
-                    calibration_metrics = calibration_report
-                    calibration_metrics['classes'] = classes_list
-                except Exception as exc:
-                    self.logger.warning(f"⚠️ Failed to compute classification calibration metrics: {exc}")
-            elif not is_classification and validation_predictions is not None:
-                try:
-                    y_array = y.values if hasattr(y, 'values') else np.asarray(y)
-                    y_float = np.asarray(y_array, dtype=float)
-                    prediction_float = np.asarray(validation_predictions, dtype=float)
-                    coverage_target = getattr(self.config, 'confidence_coverage_target', 0.9)
-                    if not (0.0 < coverage_target < 1.0):
-                        self.logger.warning(
-                            "⚠️ Invalid coverage target %.3f provided; defaulting to 0.9",
-                            coverage_target,
-                        )
-                        coverage_target = 0.9
-                    regression_report = evaluate_conformal_interval(
-                        y_float,
-                        prediction_float,
-                        coverage_target,
-                    )
-                    uncertainty_metrics = regression_report
-                    if not regression_report.get('coverage_met', True):
-                        self.logger.warning(
-                            "⚠️ Conformal interval coverage %.3f below target %.3f",
-                            regression_report.get('coverage', 0.0),
-                            regression_report.get('coverage_target', coverage_target),
-                        )
-                except Exception as exc:
-                    self.logger.warning(f"⚠️ Failed to compute regression uncertainty metrics: {exc}")
-
-        # Store feature counts
-        self.results.feature_counts = {
-            'initial': len(X.columns),
-            'stage_1': len(stage_1_features),
-            'stage_2': len(stage_2_features),
-            'stage_3': len(stage_3_features),
-            'final': len(stage_3_features)
-        }
-
-        # Store model performance
-        fold_sharpes = [float(m.get('long_short_sharpe', 0.0)) for m in fold_metrics]
-        self.results.model_performance = {
-            'final_model': final_model,
-            'model_score': model_score,
-            'fold_metrics': fold_metrics,
-            'validation_predictions': validation_predictions.to_dict(),
-            'validation_index': validation_predictions.index.tolist(),
-            'cv_scores': cv_scores.tolist(),
-            'feature_importance': dict(zip(stage_3_features, final_model.feature_importances_)),
-            'evaluation_metric': metric_name,
-            'evaluation_score': metric_value
-        }
-        if validation_predictions is not None:
-            self.results.model_performance['cv_predictions'] = (
-                validation_predictions.tolist() if hasattr(validation_predictions, 'tolist') else list(validation_predictions)
-            )
-        if validation_probabilities is not None:
-            self.results.model_performance['cv_probabilities'] = (
-                validation_probabilities.tolist() if hasattr(validation_probabilities, 'tolist') else list(validation_probabilities)
-            )
-        if classes_list is not None:
-            self.results.model_performance['cv_classes'] = classes_list
-
-        self.results.calibration_metrics = calibration_metrics
-        self.results.reliability_diagram = reliability_diagram
-        self.results.uncertainty_metrics = uncertainty_metrics
-
-        self.results.polarity_adjustments = getattr(self, 'feature_polarity_adjustments', {})
-        self.results.sign_stability = getattr(self, 'feature_sign_stability', {})
-
-    def _calculate_model_metric(self, model: Any, X: pd.DataFrame, y: pd.Series) -> Tuple[str, Optional[float]]:
-        """Calculate the final evaluation metric for the trained model."""
-
-        if self._is_classification(y):
-            return self._calculate_classification_metric(model, X, y)
-
-        try:
-            return 'r2', float(model.score(X, y))
-        except Exception:
-            return 'r2', None
-
-    def _calculate_classification_metric(self, model: Any, X: pd.DataFrame, y: pd.Series) -> Tuple[str, Optional[float]]:
-        """Calculate imbalance-aware classification metrics."""
-
-        y_array = y.values if hasattr(y, 'values') else np.asarray(y)
-        unique_classes = np.unique(y_array)
-
-        if unique_classes.size <= 1:
-            return 'average_precision', None
-
-        if unique_classes.size <= 2:
-            try:
-                if hasattr(model, 'predict_proba'):
-                    proba = model.predict_proba(X)
-                    if proba.ndim == 2 and proba.shape[1] > 1:
-                        positive_scores = proba[:, 1]
-                    else:
-                        positive_scores = proba.ravel()
-                    return 'average_precision', float(average_precision_score(y_array, positive_scores))
-                if hasattr(model, 'decision_function'):
-                    scores = model.decision_function(X)
-                    return 'average_precision', float(average_precision_score(y_array, scores))
-            except Exception as e:
-                tprint_warning(f"⚠️ Failed to calculate average precision: {e}")
-                pass
-
-        try:
-            predictions = model.predict(X)
-            return 'balanced_accuracy', float(balanced_accuracy_score(y_array, predictions))
-        except Exception:
-            return 'balanced_accuracy', None
-
-    def _handle_insufficient_features(self, X: pd.DataFrame, y: pd.Series) -> FeatureSelectionResult:
-        """Handle case where we don't have enough features."""
-        
-        self.logger.warning("⚠️ Insufficient features for multi-stage selection")
-        
-        # Use all available features
-        self.results.final_features = X.columns.tolist()
-        self.results.stage_1_features = X.columns.tolist()
-        self.results.stage_2_features = X.columns.tolist()
-        self.results.stage_3_features = X.columns.tolist()
-        
-        # Train final model
-        final_model = self._train_random_forest(X, y)
-        predictions = self._predict_signal(final_model, X)
-        trading_metrics = dict(self._evaluate_trading_metrics(y, predictions))
-        model_score = float(final_model.score(X, y)) if hasattr(final_model, 'score') else 0.0
-
-        trading_metrics.update({
-            'model_score': model_score,
-            'cv_mean_trading': trading_metrics.get('long_short_sharpe', 0.0),
-            'cv_std_trading': 0.0,
-            'cv_folds': 1,
-            'fold_metrics': [],
-            'validation_sample_count': trading_metrics.get('n_observations', len(y)),
-        })
-
-        metric_name, metric_value = self._calculate_model_metric(final_model, X, y)
-        final_scores: Dict[str, Any] = dict(trading_metrics)
-        final_scores.update({
-            'cv_metric': self.config.cv_scoring,
-            'cv_mean': float('nan'),
-            'cv_std': float('nan'),
-            'model_cv_mean': float('nan'),
-            'model_cv_std': float('nan'),
-        })
-        if metric_value is not None:
-            final_scores[metric_name] = metric_value
-
-        if final_scores.get('is_turnover_rejected'):
-            final_scores['eligible_for_selection'] = False
-        else:
-            final_scores['eligible_for_selection'] = True
-
-        self.results.final_scores = final_scores
-        self.results.eligible_for_selection = bool(final_scores.get('eligible_for_selection', True))
-        self.results.turnover_rejection_reason = final_scores.get('turnover_rejection_reason')
-
-        self.results.feature_counts = {
-            'initial': len(X.columns),
-            'stage_1': len(X.columns),
-            'stage_2': len(X.columns),
-            'stage_3': len(X.columns),
-            'final': len(X.columns)
-        }
-
-        self.results.model_performance = {
-            'final_model': final_model,
-            'model_score': model_score,
-            'cv_scores': [trading_metrics.get('long_short_sharpe', 0.0)],
-            'fold_metrics': [],
-            'validation_predictions': predictions.to_dict(),
-            'validation_index': predictions.index.tolist(),
-            'feature_importance': dict(zip(X.columns, getattr(final_model, 'feature_importances_', np.zeros(len(X.columns)))))
-        }
-
-        self.results.polarity_adjustments = getattr(self, 'feature_polarity_adjustments', {})
-        self.results.sign_stability = getattr(self, 'feature_sign_stability', {})
-
-        return self.results
-    
-    def _save_analysis(self):
-        """Save analysis results."""
-        tprint_debug("💾 Saving analysis results")
-        try:
-            from datetime import datetime
-
-            output_dir = Path(self.config.output_directory)
-            output_dir.mkdir(parents=True, exist_ok=True)
-
-            # Save results summary with proper outcomes naming convention
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            results_file = output_dir / f"market_analysis_feature_selection_outcome_{timestamp}.json"
-            
-            # Convert results to serializable format
-            results_dict = {
-                'feature_counts': self.results.feature_counts,
-                'final_features': self.results.final_features,
-                'scores': {
-                    'stage_1': self.results.stage_1_scores,
-                    'stage_2': self.results.stage_2_scores,
-                    'stage_3': self.results.stage_3_scores,
-                    'final': self.results.final_scores
-                },
-                'calibration': self.results.calibration_metrics,
-                'uncertainty': self.results.uncertainty_metrics,
-                'reliability_diagram': self.results.reliability_diagram,
-                'entropy_filter': {
-                    'removed_features': self.results.entropy_removed_features,
-                    'stability_scores': self.results.entropy_stability,
-                    'variance': self.results.entropy_variance,
-                    'config': {
-                        'num_slices': self.config.entropy_num_slices,
-                        'min_slice_size': self.config.entropy_min_slice_size,
-                        'variance_threshold': self.config.entropy_variance_threshold,
-                        'max_bins': self.config.entropy_max_bins,
-                        'min_unique_values': self.config.entropy_min_unique_values,
-                    },
-                },
-                'polarity_adjustments': getattr(self.results, 'polarity_adjustments', {}),
-                'sign_stability': getattr(self.results, 'sign_stability', {}),
-                'selection_time': self.results.selection_time,
-                'config': {
-                    'initial_features': self.config.initial_features,
-                    'stage_1_target': self.config.stage_1_target,
-                    'stage_2_target': self.config.stage_2_target,
-                    'stage_3_target': self.config.stage_3_target,
-                    'rf_n_estimators': self.config.rf_n_estimators,
-                    'cv_folds': self.config.cv_folds,
-                    'enable_uncertainty_reporting': self.config.enable_uncertainty_reporting,
-                    'reliability_bins': self.config.reliability_bins,
-                    'confidence_coverage_target': self.config.confidence_coverage_target,
-                }
-            }
-            
-            with open(results_file, 'w') as f:
-                json.dump(results_dict, f, indent=2)
-            
-            self.logger.info(f"💾 Analysis results saved to {results_file}")
-            
-            # Save final model if requested
-            if self.config.save_models and hasattr(self.results, 'model_performance'):
-                model_file = output_dir / f"market_analysis_feature_selection_model_{timestamp}.joblib"
-                joblib.dump(self.results.model_performance['final_model'], model_file)
-                self.logger.info(f"💾 Final model saved to {model_file}")
-                
-        except Exception as e:
-            self.logger.error(f"❌ Failed to save analysis: {e}")
-
-# Convenience functions
-def run_final_feature_selection(X: pd.DataFrame,
-                               y: pd.Series,
-                               config: Optional[FeatureSelectionConfig] = None) -> FeatureSelectionResult:
-    """Run final feature selection pipeline."""
-    tprint("🚀 Running final feature selection helper")
-    selector = MultiStageFeatureSelector(config)
-    tprint("📊 Executing multi-stage selector")
-    return selector.select_features(X, y)
-
-def get_final_features(X: pd.DataFrame,
-                      y: pd.Series,
-                      target_count: int = 60,
-                      config: Optional[FeatureSelectionConfig] = None) -> List[str]:
-    """Get final selected features."""
-    if config is None:
-        config = FeatureSelectionConfig()
-        config.stage_3_target = target_count
-        tprint(f"🎯 Configuring stage 3 target for final features: {target_count}")
-
-    result = run_final_feature_selection(X, y, config)
-    tprint(f"✅ Final feature count: {len(result.final_features)}")
-    return result.final_features
